MutCount concat method re-written : functional tests ok + minor improvement in separated method
