+++ galaxy_wrappers/07_MutCount/scripts/S01a_codons_counting.py
@@ -1,0 +1,1 @@
+../../../scripts/07_MutCount/S01a_codons_counting.py
