--- galaxy_wrappers/07_MutCount/scripts/S01a_mutcount_pairs.py
@@ -1,1 +1,0 @@
-../../../scripts/07_MutCount/S01a_mutcount_pairs.py
