--- galaxy_wrappers/07_MutCount/scripts/S02a_codon_counting.py
@@ -1,1 +1,0 @@
-../../../scripts/07_MutCount/S02a_codon_counting.py
