+++ galaxy_wrappers/07_MutCount/test-data/OUT_aa/AVLIMFYW.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_count_AVLIMFYW,Ac_prop_AVLIMFYW,Ac_count_AVLIM,Ac_prop_AVLIM,Ac_count_FYW,Ac_prop_FYW,Pu_count_AVLIMFYW,Pu_prop_AVLIMFYW,Pu_count_AVLIM,Pu_prop_AVLIM,Pu_count_FYW,Pu_prop_FYW,Am_count_AVLIMFYW,Am_prop_AVLIMFYW,Am_count_AVLIM,Am_prop_AVLIM,Am_count_FYW,Am_prop_FYW,Ap_count_AVLIMFYW,Ap_prop_AVLIMFYW,Ap_count_AVLIM,Ap_prop_AVLIM,Ap_count_FYW,Ap_prop_FYW,Pf_count_AVLIMFYW,Pf_prop_AVLIMFYW,Pf_count_AVLIM,Pf_prop_AVLIM,Pf_count_FYW,Pf_prop_FYW,Pg_count_AVLIMFYW,Pg_prop_AVLIMFYW,Pg_count_AVLIM,Pg_prop_AVLIM,Pg_count_FYW,Pg_prop_FYW,Ph_count_AVLIMFYW,Ph_prop_AVLIMFYW,Ph_count_AVLIM,Ph_prop_AVLIM,Ph_count_FYW,Ph_prop_FYW,Ps_count_AVLIMFYW,Ps_prop_AVLIMFYW,Ps_count_AVLIM,Ps_prop_AVLIM,Ps_count_FYW,Ps_prop_FYW,Pp_count_AVLIMFYW,Pp_prop_AVLIMFYW,Pp_count_AVLIM,Pp_prop_AVLIM,Pp_count_FYW,Pp_prop_FYW,Pa_count_AVLIMFYW,Pa_prop_AVLIMFYW,Pa_count_AVLIM,Pa_prop_AVLIM,Pa_count_FYW,Pa_prop_FYW,
-locus2_sp6_sp6.fasta,27.00000,0.42188,21.00000,0.32812,6.00000,0.09375,28.00000,0.43750,22.00000,0.34375,6.00000,0.09375,NA,NA,NA,NA,NA,NA,28.00000,0.43750,22.00000,0.34375,6.00000,0.09375,NA,NA,NA,NA,NA,NA,28.00000,0.43750,22.00000,0.34375,6.00000,0.09375,30.00000,0.46875,24.00000,0.37500,6.00000,0.09375,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,28.00000,0.43750,22.00000,0.34375,6.00000,0.09375,
-locus1_sp6_sp6.fasta,25.00000,0.32051,22.00000,0.28205,3.00000,0.03846,25.00000,0.32051,23.00000,0.29487,2.00000,0.02564,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,25.00000,0.32051,23.00000,0.29487,2.00000,0.02564,24.00000,0.30769,22.00000,0.28205,2.00000,0.02564,24.00000,0.30769,22.00000,0.28205,2.00000,0.02564,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,24.00000,0.30769,22.00000,0.28205,2.00000,0.02564,
-locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,28.00000,0.36364,25.00000,0.32468,3.00000,0.03896,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,29.00000,0.37662,26.00000,0.33766,3.00000,0.03896,30.00000,0.38961,27.00000,0.35065,3.00000,0.03896,28.00000,0.36364,25.00000,0.32468,3.00000,0.03896,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,24.00000,0.36364,22.00000,0.33333,2.00000,0.03030,29.00000,0.37662,26.00000,0.33766,3.00000,0.03896,
+LOCUS,Ac_count_AVLIMFYW,Ac_prop_AVLIMFYW,Ac_count_AVLIM,Ac_prop_AVLIM,Ac_count_FYW,Ac_prop_FYW,Pu_count_AVLIMFYW,Pu_prop_AVLIMFYW,Pu_count_AVLIM,Pu_prop_AVLIM,Pu_count_FYW,Pu_prop_FYW,Am_count_AVLIMFYW,Am_prop_AVLIMFYW,Am_count_AVLIM,Am_prop_AVLIM,Am_count_FYW,Am_prop_FYW,Ap_count_AVLIMFYW,Ap_prop_AVLIMFYW,Ap_count_AVLIM,Ap_prop_AVLIM,Ap_count_FYW,Ap_prop_FYW,Pf_count_AVLIMFYW,Pf_prop_AVLIMFYW,Pf_count_AVLIM,Pf_prop_AVLIM,Pf_count_FYW,Pf_prop_FYW,Pg_count_AVLIMFYW,Pg_prop_AVLIMFYW,Pg_count_AVLIM,Pg_prop_AVLIM,Pg_count_FYW,Pg_prop_FYW,Ph_count_AVLIMFYW,Ph_prop_AVLIMFYW,Ph_count_AVLIM,Ph_prop_AVLIM,Ph_count_FYW,Ph_prop_FYW,Ps_count_AVLIMFYW,Ps_prop_AVLIMFYW,Ps_count_AVLIM,Ps_prop_AVLIM,Ps_count_FYW,Ps_prop_FYW,Pp_count_AVLIMFYW,Pp_prop_AVLIMFYW,Pp_count_AVLIM,Pp_prop_AVLIM,Pp_count_FYW,Pp_prop_FYW,Pa_count_AVLIMFYW,Pa_prop_AVLIMFYW,Pa_count_AVLIM,Pa_prop_AVLIM,Pa_count_FYW,Pa_prop_FYW
+locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,28.00000,0.36364,25.00000,0.32468,3.00000,0.03896,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,29.00000,0.37662,26.00000,0.33766,3.00000,0.03896,30.00000,0.38961,27.00000,0.35065,3.00000,0.03896,28.00000,0.36364,25.00000,0.32468,3.00000,0.03896,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,24.00000,0.36364,22.00000,0.33333,2.00000,0.03030,29.00000,0.37662,26.00000,0.33766,3.00000,0.03896
+locus1_sp6_sp6.fasta,25.00000,0.32051,22.00000,0.28205,3.00000,0.03846,25.00000,0.32051,23.00000,0.29487,2.00000,0.02564,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,25.00000,0.32051,23.00000,0.29487,2.00000,0.02564,24.00000,0.30769,22.00000,0.28205,2.00000,0.02564,24.00000,0.30769,22.00000,0.28205,2.00000,0.02564,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,24.00000,0.30769,22.00000,0.28205,2.00000,0.02564
+locus2_sp6_sp6.fasta,27.00000,0.42188,21.00000,0.32812,6.00000,0.09375,28.00000,0.43750,22.00000,0.34375,6.00000,0.09375,NA,NA,NA,NA,NA,NA,28.00000,0.43750,22.00000,0.34375,6.00000,0.09375,NA,NA,NA,NA,NA,NA,28.00000,0.43750,22.00000,0.34375,6.00000,0.09375,30.00000,0.46875,24.00000,0.37500,6.00000,0.09375,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,28.00000,0.43750,22.00000,0.34375,6.00000,0.09375
