+++ galaxy_wrappers/07_MutCount/test-data/OUT_aa/EK_QH.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_count_EK,Ac_prop_EK,Ac_count_QH,Ac_prop_QH,Ac_ratio_EK_vs_QH,Pu_count_EK,Pu_prop_EK,Pu_count_QH,Pu_prop_QH,Pu_ratio_EK_vs_QH,Am_count_EK,Am_prop_EK,Am_count_QH,Am_prop_QH,Am_ratio_EK_vs_QH,Ap_count_EK,Ap_prop_EK,Ap_count_QH,Ap_prop_QH,Ap_ratio_EK_vs_QH,Pf_count_EK,Pf_prop_EK,Pf_count_QH,Pf_prop_QH,Pf_ratio_EK_vs_QH,Pg_count_EK,Pg_prop_EK,Pg_count_QH,Pg_prop_QH,Pg_ratio_EK_vs_QH,Ph_count_EK,Ph_prop_EK,Ph_count_QH,Ph_prop_QH,Ph_ratio_EK_vs_QH,Ps_count_EK,Ps_prop_EK,Ps_count_QH,Ps_prop_QH,Ps_ratio_EK_vs_QH,Pp_count_EK,Pp_prop_EK,Pp_count_QH,Pp_prop_QH,Pp_ratio_EK_vs_QH,Pa_count_EK,Pa_prop_EK,Pa_count_QH,Pa_prop_QH,Pa_ratio_EK_vs_QH,
-locus2_sp6_sp6.fasta,13.00000,0.20312,2.00000,0.03125,6.50000,13.00000,0.20312,2.00000,0.03125,6.50000,NA,NA,NA,NA,NA,13.00000,0.20312,2.00000,0.03125,6.50000,NA,NA,NA,NA,NA,14.00000,0.21875,2.00000,0.03125,7.00000,13.00000,0.20312,2.00000,0.03125,6.50000,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,14.00000,0.21875,2.00000,0.03125,7.00000,
-locus1_sp6_sp6.fasta,10.00000,0.12821,11.00000,0.14103,0.90909,11.00000,0.14103,11.00000,0.14103,1.00000,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,10.00000,0.12821,11.00000,0.14103,0.90909,10.00000,0.12821,11.00000,0.14103,0.90909,10.00000,0.12821,11.00000,0.14103,0.90909,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,10.00000,0.12821,11.00000,0.14103,0.90909,
-locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,0.00000,10.00000,0.12987,5.00000,0.06494,2.00000,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,9.00000,0.11688,4.00000,0.05195,2.25000,9.00000,0.11688,5.00000,0.06494,1.80000,10.00000,0.12987,4.00000,0.05195,2.50000,0.00000,0.00000,0.00000,0.00000,0.00000,9.00000,0.13636,4.00000,0.06061,2.25000,9.00000,0.11688,5.00000,0.06494,1.80000,
+LOCUS,Ac_count_EK,Ac_prop_EK,Ac_count_QH,Ac_prop_QH,Ac_ratio_EK_vs_QH,Pu_count_EK,Pu_prop_EK,Pu_count_QH,Pu_prop_QH,Pu_ratio_EK_vs_QH,Am_count_EK,Am_prop_EK,Am_count_QH,Am_prop_QH,Am_ratio_EK_vs_QH,Ap_count_EK,Ap_prop_EK,Ap_count_QH,Ap_prop_QH,Ap_ratio_EK_vs_QH,Pf_count_EK,Pf_prop_EK,Pf_count_QH,Pf_prop_QH,Pf_ratio_EK_vs_QH,Pg_count_EK,Pg_prop_EK,Pg_count_QH,Pg_prop_QH,Pg_ratio_EK_vs_QH,Ph_count_EK,Ph_prop_EK,Ph_count_QH,Ph_prop_QH,Ph_ratio_EK_vs_QH,Ps_count_EK,Ps_prop_EK,Ps_count_QH,Ps_prop_QH,Ps_ratio_EK_vs_QH,Pp_count_EK,Pp_prop_EK,Pp_count_QH,Pp_prop_QH,Pp_ratio_EK_vs_QH,Pa_count_EK,Pa_prop_EK,Pa_count_QH,Pa_prop_QH,Pa_ratio_EK_vs_QH
+locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,0.00000,10.00000,0.12987,5.00000,0.06494,2.00000,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,9.00000,0.11688,4.00000,0.05195,2.25000,9.00000,0.11688,5.00000,0.06494,1.80000,10.00000,0.12987,4.00000,0.05195,2.50000,0.00000,0.00000,0.00000,0.00000,0.00000,9.00000,0.13636,4.00000,0.06061,2.25000,9.00000,0.11688,5.00000,0.06494,1.80000
+locus1_sp6_sp6.fasta,10.00000,0.12821,11.00000,0.14103,0.90909,11.00000,0.14103,11.00000,0.14103,1.00000,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,10.00000,0.12821,11.00000,0.14103,0.90909,10.00000,0.12821,11.00000,0.14103,0.90909,10.00000,0.12821,11.00000,0.14103,0.90909,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,10.00000,0.12821,11.00000,0.14103,0.90909
+locus2_sp6_sp6.fasta,13.00000,0.20312,2.00000,0.03125,6.50000,13.00000,0.20312,2.00000,0.03125,6.50000,NA,NA,NA,NA,NA,13.00000,0.20312,2.00000,0.03125,6.50000,NA,NA,NA,NA,NA,14.00000,0.21875,2.00000,0.03125,7.00000,13.00000,0.20312,2.00000,0.03125,6.50000,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,14.00000,0.21875,2.00000,0.03125,7.00000
