+++ galaxy_wrappers/07_MutCount/test-data/OUT_aa/ERK_DNQTSH.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_count_ERK,Ac_prop_ERK,Ac_count_DNQTSH,Ac_prop_DNQTSH,Ac_ratio_ERK_vs_DNQTSH,Pu_count_ERK,Pu_prop_ERK,Pu_count_DNQTSH,Pu_prop_DNQTSH,Pu_ratio_ERK_vs_DNQTSH,Am_count_ERK,Am_prop_ERK,Am_count_DNQTSH,Am_prop_DNQTSH,Am_ratio_ERK_vs_DNQTSH,Ap_count_ERK,Ap_prop_ERK,Ap_count_DNQTSH,Ap_prop_DNQTSH,Ap_ratio_ERK_vs_DNQTSH,Pf_count_ERK,Pf_prop_ERK,Pf_count_DNQTSH,Pf_prop_DNQTSH,Pf_ratio_ERK_vs_DNQTSH,Pg_count_ERK,Pg_prop_ERK,Pg_count_DNQTSH,Pg_prop_DNQTSH,Pg_ratio_ERK_vs_DNQTSH,Ph_count_ERK,Ph_prop_ERK,Ph_count_DNQTSH,Ph_prop_DNQTSH,Ph_ratio_ERK_vs_DNQTSH,Ps_count_ERK,Ps_prop_ERK,Ps_count_DNQTSH,Ps_prop_DNQTSH,Ps_ratio_ERK_vs_DNQTSH,Pp_count_ERK,Pp_prop_ERK,Pp_count_DNQTSH,Pp_prop_DNQTSH,Pp_ratio_ERK_vs_DNQTSH,Pa_count_ERK,Pa_prop_ERK,Pa_count_DNQTSH,Pa_prop_DNQTSH,Pa_ratio_ERK_vs_DNQTSH,
-locus2_sp6_sp6.fasta,13.00000,0.20312,14.00000,0.21875,0.92857,13.00000,0.20312,14.00000,0.21875,0.92857,NA,NA,NA,NA,NA,13.00000,0.20312,13.00000,0.20312,1.00000,NA,NA,NA,NA,NA,14.00000,0.21875,13.00000,0.20312,1.07692,13.00000,0.20312,14.00000,0.21875,0.92857,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,14.00000,0.21875,13.00000,0.20312,1.07692,
-locus1_sp6_sp6.fasta,16.00000,0.20513,31.00000,0.39744,0.51613,18.00000,0.23077,27.00000,0.34615,0.66667,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,18.00000,0.23077,30.00000,0.38462,0.60000,17.00000,0.21795,31.00000,0.39744,0.54839,17.00000,0.21795,31.00000,0.39744,0.54839,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,17.00000,0.21795,31.00000,0.39744,0.54839,
-locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,0.00000,11.00000,0.14286,31.00000,0.40260,0.35484,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,11.00000,0.14286,32.00000,0.41558,0.34375,10.00000,0.12987,30.00000,0.38961,0.33333,11.00000,0.14286,30.00000,0.38961,0.36667,0.00000,0.00000,0.00000,0.00000,0.00000,10.00000,0.15152,27.00000,0.40909,0.37037,10.00000,0.12987,31.00000,0.40260,0.32258,
+LOCUS,Ac_count_ERK,Ac_prop_ERK,Ac_count_DNQTSH,Ac_prop_DNQTSH,Ac_ratio_ERK_vs_DNQTSH,Pu_count_ERK,Pu_prop_ERK,Pu_count_DNQTSH,Pu_prop_DNQTSH,Pu_ratio_ERK_vs_DNQTSH,Am_count_ERK,Am_prop_ERK,Am_count_DNQTSH,Am_prop_DNQTSH,Am_ratio_ERK_vs_DNQTSH,Ap_count_ERK,Ap_prop_ERK,Ap_count_DNQTSH,Ap_prop_DNQTSH,Ap_ratio_ERK_vs_DNQTSH,Pf_count_ERK,Pf_prop_ERK,Pf_count_DNQTSH,Pf_prop_DNQTSH,Pf_ratio_ERK_vs_DNQTSH,Pg_count_ERK,Pg_prop_ERK,Pg_count_DNQTSH,Pg_prop_DNQTSH,Pg_ratio_ERK_vs_DNQTSH,Ph_count_ERK,Ph_prop_ERK,Ph_count_DNQTSH,Ph_prop_DNQTSH,Ph_ratio_ERK_vs_DNQTSH,Ps_count_ERK,Ps_prop_ERK,Ps_count_DNQTSH,Ps_prop_DNQTSH,Ps_ratio_ERK_vs_DNQTSH,Pp_count_ERK,Pp_prop_ERK,Pp_count_DNQTSH,Pp_prop_DNQTSH,Pp_ratio_ERK_vs_DNQTSH,Pa_count_ERK,Pa_prop_ERK,Pa_count_DNQTSH,Pa_prop_DNQTSH,Pa_ratio_ERK_vs_DNQTSH
+locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,0.00000,11.00000,0.14286,31.00000,0.40260,0.35484,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,11.00000,0.14286,32.00000,0.41558,0.34375,10.00000,0.12987,30.00000,0.38961,0.33333,11.00000,0.14286,30.00000,0.38961,0.36667,0.00000,0.00000,0.00000,0.00000,0.00000,10.00000,0.15152,27.00000,0.40909,0.37037,10.00000,0.12987,31.00000,0.40260,0.32258
+locus1_sp6_sp6.fasta,16.00000,0.20513,31.00000,0.39744,0.51613,18.00000,0.23077,27.00000,0.34615,0.66667,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,18.00000,0.23077,30.00000,0.38462,0.60000,17.00000,0.21795,31.00000,0.39744,0.54839,17.00000,0.21795,31.00000,0.39744,0.54839,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,17.00000,0.21795,31.00000,0.39744,0.54839
+locus2_sp6_sp6.fasta,13.00000,0.20312,14.00000,0.21875,0.92857,13.00000,0.20312,14.00000,0.21875,0.92857,NA,NA,NA,NA,NA,13.00000,0.20312,13.00000,0.20312,1.00000,NA,NA,NA,NA,NA,14.00000,0.21875,13.00000,0.20312,1.07692,13.00000,0.20312,14.00000,0.21875,0.92857,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,14.00000,0.21875,13.00000,0.20312,1.07692
