+++ galaxy_wrappers/07_MutCount/test-data/OUT_aa/FYMINK_GARP.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_count_FYMINK,Ac_prop_FYMINK,Ac_count_GARP,Ac_prop_GARP,Pu_count_FYMINK,Pu_prop_FYMINK,Pu_count_GARP,Pu_prop_GARP,Am_count_FYMINK,Am_prop_FYMINK,Am_count_GARP,Am_prop_GARP,Ap_count_FYMINK,Ap_prop_FYMINK,Ap_count_GARP,Ap_prop_GARP,Pf_count_FYMINK,Pf_prop_FYMINK,Pf_count_GARP,Pf_prop_GARP,Pg_count_FYMINK,Pg_prop_FYMINK,Pg_count_GARP,Pg_prop_GARP,Ph_count_FYMINK,Ph_prop_FYMINK,Ph_count_GARP,Ph_prop_GARP,Ps_count_FYMINK,Ps_prop_FYMINK,Ps_count_GARP,Ps_prop_GARP,Pp_count_FYMINK,Pp_prop_FYMINK,Pp_count_GARP,Pp_prop_GARP,Pa_count_FYMINK,Pa_prop_FYMINK,Pa_count_GARP,Pa_prop_GARP,
-locus2_sp6_sp6.fasta,21.00000,0.32812,16.00000,0.25000,20.00000,0.31250,15.00000,0.23438,NA,NA,NA,NA,21.00000,0.32812,16.00000,0.25000,NA,NA,NA,NA,21.00000,0.32812,15.00000,0.23438,20.00000,0.31250,15.00000,0.23438,NA,NA,NA,NA,NA,NA,NA,NA,21.00000,0.32812,15.00000,0.23438,
-locus1_sp6_sp6.fasta,15.00000,0.19231,17.00000,0.21795,12.00000,0.15385,20.00000,0.25641,NA,NA,NA,NA,NA,NA,NA,NA,13.00000,0.16667,19.00000,0.24359,13.00000,0.16667,18.00000,0.23077,12.00000,0.15385,18.00000,0.23077,NA,NA,NA,NA,NA,NA,NA,NA,13.00000,0.16667,18.00000,0.23077,
-locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,16.00000,0.20779,12.00000,0.15584,NA,NA,NA,NA,NA,NA,NA,NA,15.00000,0.19481,13.00000,0.16883,12.00000,0.15584,15.00000,0.19481,17.00000,0.22078,14.00000,0.18182,0.00000,0.00000,0.00000,0.00000,12.00000,0.18182,8.00000,0.12121,12.00000,0.15584,14.00000,0.18182,
+LOCUS,Ac_count_FYMINK,Ac_prop_FYMINK,Ac_count_GARP,Ac_prop_GARP,Pu_count_FYMINK,Pu_prop_FYMINK,Pu_count_GARP,Pu_prop_GARP,Am_count_FYMINK,Am_prop_FYMINK,Am_count_GARP,Am_prop_GARP,Ap_count_FYMINK,Ap_prop_FYMINK,Ap_count_GARP,Ap_prop_GARP,Pf_count_FYMINK,Pf_prop_FYMINK,Pf_count_GARP,Pf_prop_GARP,Pg_count_FYMINK,Pg_prop_FYMINK,Pg_count_GARP,Pg_prop_GARP,Ph_count_FYMINK,Ph_prop_FYMINK,Ph_count_GARP,Ph_prop_GARP,Ps_count_FYMINK,Ps_prop_FYMINK,Ps_count_GARP,Ps_prop_GARP,Pp_count_FYMINK,Pp_prop_FYMINK,Pp_count_GARP,Pp_prop_GARP,Pa_count_FYMINK,Pa_prop_FYMINK,Pa_count_GARP,Pa_prop_GARP
+locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,16.00000,0.20779,12.00000,0.15584,NA,NA,NA,NA,NA,NA,NA,NA,15.00000,0.19481,13.00000,0.16883,12.00000,0.15584,15.00000,0.19481,17.00000,0.22078,14.00000,0.18182,0.00000,0.00000,0.00000,0.00000,12.00000,0.18182,8.00000,0.12121,12.00000,0.15584,14.00000,0.18182
+locus1_sp6_sp6.fasta,15.00000,0.19231,17.00000,0.21795,12.00000,0.15385,20.00000,0.25641,NA,NA,NA,NA,NA,NA,NA,NA,13.00000,0.16667,19.00000,0.24359,13.00000,0.16667,18.00000,0.23077,12.00000,0.15385,18.00000,0.23077,NA,NA,NA,NA,NA,NA,NA,NA,13.00000,0.16667,18.00000,0.23077
+locus2_sp6_sp6.fasta,21.00000,0.32812,16.00000,0.25000,20.00000,0.31250,15.00000,0.23438,NA,NA,NA,NA,21.00000,0.32812,16.00000,0.25000,NA,NA,NA,NA,21.00000,0.32812,15.00000,0.23438,20.00000,0.31250,15.00000,0.23438,NA,NA,NA,NA,NA,NA,NA,NA,21.00000,0.32812,15.00000,0.23438
