+++ galaxy_wrappers/07_MutCount/test-data/OUT_aa/IVYWREL.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_count_IVYWREL,Ac_prop_IVYWREL,Pu_count_IVYWREL,Pu_prop_IVYWREL,Am_count_IVYWREL,Am_prop_IVYWREL,Ap_count_IVYWREL,Ap_prop_IVYWREL,Pf_count_IVYWREL,Pf_prop_IVYWREL,Pg_count_IVYWREL,Pg_prop_IVYWREL,Ph_count_IVYWREL,Ph_prop_IVYWREL,Ps_count_IVYWREL,Ps_prop_IVYWREL,Pp_count_IVYWREL,Pp_prop_IVYWREL,Pa_count_IVYWREL,Pa_prop_IVYWREL,
-locus2_sp6_sp6.fasta,21.00000,0.32812,23.00000,0.35938,NA,NA,23.00000,0.35938,NA,NA,22.00000,0.34375,23.00000,0.35938,NA,NA,NA,NA,22.00000,0.34375,
-locus1_sp6_sp6.fasta,29.00000,0.37179,31.00000,0.39744,NA,NA,NA,NA,29.00000,0.37179,28.00000,0.35897,29.00000,0.37179,NA,NA,NA,NA,28.00000,0.35897,
-locus1_sp8_sp8.fasta,0.00000,0.00000,28.00000,0.36364,NA,NA,NA,NA,27.00000,0.35065,27.00000,0.35065,28.00000,0.36364,0.00000,0.00000,27.00000,0.40909,27.00000,0.35065,
+LOCUS,Ac_count_IVYWREL,Ac_prop_IVYWREL,Pu_count_IVYWREL,Pu_prop_IVYWREL,Am_count_IVYWREL,Am_prop_IVYWREL,Ap_count_IVYWREL,Ap_prop_IVYWREL,Pf_count_IVYWREL,Pf_prop_IVYWREL,Pg_count_IVYWREL,Pg_prop_IVYWREL,Ph_count_IVYWREL,Ph_prop_IVYWREL,Ps_count_IVYWREL,Ps_prop_IVYWREL,Pp_count_IVYWREL,Pp_prop_IVYWREL,Pa_count_IVYWREL,Pa_prop_IVYWREL
+locus1_sp8_sp8.fasta,0.00000,0.00000,28.00000,0.36364,NA,NA,NA,NA,27.00000,0.35065,27.00000,0.35065,28.00000,0.36364,0.00000,0.00000,27.00000,0.40909,27.00000,0.35065
+locus1_sp6_sp6.fasta,29.00000,0.37179,31.00000,0.39744,NA,NA,NA,NA,29.00000,0.37179,28.00000,0.35897,29.00000,0.37179,NA,NA,NA,NA,28.00000,0.35897
+locus2_sp6_sp6.fasta,21.00000,0.32812,23.00000,0.35938,NA,NA,23.00000,0.35938,NA,NA,22.00000,0.34375,23.00000,0.35938,NA,NA,NA,NA,22.00000,0.34375
