+++ galaxy_wrappers/07_MutCount/test-data/OUT_aa/PAYRE-MVGDS.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_count_PAYRE,Ac_prop_PAYRE,Ac_count_AC,Ac_prop_AC,Ac_count_MVGDS,Ac_prop_MVGDS,Ac_ratio_PAYRE_vs_MVGDS,Ac_ratio_AC_vs_MVGDS,Pu_count_PAYRE,Pu_prop_PAYRE,Pu_count_AC,Pu_prop_AC,Pu_count_MVGDS,Pu_prop_MVGDS,Pu_ratio_PAYRE_vs_MVGDS,Pu_ratio_AC_vs_MVGDS,Am_count_PAYRE,Am_prop_PAYRE,Am_count_AC,Am_prop_AC,Am_count_MVGDS,Am_prop_MVGDS,Am_ratio_PAYRE_vs_MVGDS,Am_ratio_AC_vs_MVGDS,Ap_count_PAYRE,Ap_prop_PAYRE,Ap_count_AC,Ap_prop_AC,Ap_count_MVGDS,Ap_prop_MVGDS,Ap_ratio_PAYRE_vs_MVGDS,Ap_ratio_AC_vs_MVGDS,Pf_count_PAYRE,Pf_prop_PAYRE,Pf_count_AC,Pf_prop_AC,Pf_count_MVGDS,Pf_prop_MVGDS,Pf_ratio_PAYRE_vs_MVGDS,Pf_ratio_AC_vs_MVGDS,Pg_count_PAYRE,Pg_prop_PAYRE,Pg_count_AC,Pg_prop_AC,Pg_count_MVGDS,Pg_prop_MVGDS,Pg_ratio_PAYRE_vs_MVGDS,Pg_ratio_AC_vs_MVGDS,Ph_count_PAYRE,Ph_prop_PAYRE,Ph_count_AC,Ph_prop_AC,Ph_count_MVGDS,Ph_prop_MVGDS,Ph_ratio_PAYRE_vs_MVGDS,Ph_ratio_AC_vs_MVGDS,Ps_count_PAYRE,Ps_prop_PAYRE,Ps_count_AC,Ps_prop_AC,Ps_count_MVGDS,Ps_prop_MVGDS,Ps_ratio_PAYRE_vs_MVGDS,Ps_ratio_AC_vs_MVGDS,Pp_count_PAYRE,Pp_prop_PAYRE,Pp_count_AC,Pp_prop_AC,Pp_count_MVGDS,Pp_prop_MVGDS,Pp_ratio_PAYRE_vs_MVGDS,Pp_ratio_AC_vs_MVGDS,Pa_count_PAYRE,Pa_prop_PAYRE,Pa_count_AC,Pa_prop_AC,Pa_count_MVGDS,Pa_prop_MVGDS,Pa_ratio_PAYRE_vs_MVGDS,Pa_ratio_AC_vs_MVGDS,
-locus2_sp6_sp6.fasta,18.00000,0.28125,6.00000,0.09375,20.00000,0.31250,0.90000,0.30000,18.00000,0.28125,6.00000,0.09375,21.00000,0.32812,0.85714,0.28571,NA,NA,NA,NA,NA,NA,NA,NA,18.00000,0.28125,6.00000,0.09375,20.00000,0.31250,0.90000,0.30000,NA,NA,NA,NA,NA,NA,NA,NA,17.00000,0.26562,6.00000,0.09375,20.00000,0.31250,0.85000,0.30000,20.00000,0.31250,8.00000,0.12500,19.00000,0.29688,1.05263,0.42105,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,17.00000,0.26562,6.00000,0.09375,20.00000,0.31250,0.85000,0.30000,
-locus1_sp6_sp6.fasta,24.00000,0.30769,7.00000,0.08974,20.00000,0.25641,1.20000,0.35000,25.00000,0.32051,7.00000,0.08974,20.00000,0.25641,1.25000,0.35000,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,25.00000,0.32051,8.00000,0.10256,21.00000,0.26923,1.19048,0.38095,24.00000,0.30769,7.00000,0.08974,22.00000,0.28205,1.09091,0.31818,24.00000,0.30769,7.00000,0.08974,20.00000,0.25641,1.20000,0.35000,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,24.00000,0.30769,7.00000,0.08974,22.00000,0.28205,1.09091,0.31818,
-locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,18.00000,0.23377,4.00000,0.05195,29.00000,0.37662,0.62069,0.13793,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,19.00000,0.24675,6.00000,0.07792,30.00000,0.38961,0.63333,0.20000,20.00000,0.25974,7.00000,0.09091,32.00000,0.41558,0.62500,0.21875,20.00000,0.25974,5.00000,0.06494,29.00000,0.37662,0.68966,0.17241,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,14.00000,0.21212,2.00000,0.03030,26.00000,0.39394,0.53846,0.07692,19.00000,0.24675,6.00000,0.07792,32.00000,0.41558,0.59375,0.18750,
+LOCUS,Ac_count_PAYRE,Ac_prop_PAYRE,Ac_count_AC,Ac_prop_AC,Ac_count_MVGDS,Ac_prop_MVGDS,Ac_ratio_PAYRE_vs_MVGDS,Ac_ratio_AC_vs_MVGDS,Pu_count_PAYRE,Pu_prop_PAYRE,Pu_count_AC,Pu_prop_AC,Pu_count_MVGDS,Pu_prop_MVGDS,Pu_ratio_PAYRE_vs_MVGDS,Pu_ratio_AC_vs_MVGDS,Am_count_PAYRE,Am_prop_PAYRE,Am_count_AC,Am_prop_AC,Am_count_MVGDS,Am_prop_MVGDS,Am_ratio_PAYRE_vs_MVGDS,Am_ratio_AC_vs_MVGDS,Ap_count_PAYRE,Ap_prop_PAYRE,Ap_count_AC,Ap_prop_AC,Ap_count_MVGDS,Ap_prop_MVGDS,Ap_ratio_PAYRE_vs_MVGDS,Ap_ratio_AC_vs_MVGDS,Pf_count_PAYRE,Pf_prop_PAYRE,Pf_count_AC,Pf_prop_AC,Pf_count_MVGDS,Pf_prop_MVGDS,Pf_ratio_PAYRE_vs_MVGDS,Pf_ratio_AC_vs_MVGDS,Pg_count_PAYRE,Pg_prop_PAYRE,Pg_count_AC,Pg_prop_AC,Pg_count_MVGDS,Pg_prop_MVGDS,Pg_ratio_PAYRE_vs_MVGDS,Pg_ratio_AC_vs_MVGDS,Ph_count_PAYRE,Ph_prop_PAYRE,Ph_count_AC,Ph_prop_AC,Ph_count_MVGDS,Ph_prop_MVGDS,Ph_ratio_PAYRE_vs_MVGDS,Ph_ratio_AC_vs_MVGDS,Ps_count_PAYRE,Ps_prop_PAYRE,Ps_count_AC,Ps_prop_AC,Ps_count_MVGDS,Ps_prop_MVGDS,Ps_ratio_PAYRE_vs_MVGDS,Ps_ratio_AC_vs_MVGDS,Pp_count_PAYRE,Pp_prop_PAYRE,Pp_count_AC,Pp_prop_AC,Pp_count_MVGDS,Pp_prop_MVGDS,Pp_ratio_PAYRE_vs_MVGDS,Pp_ratio_AC_vs_MVGDS,Pa_count_PAYRE,Pa_prop_PAYRE,Pa_count_AC,Pa_prop_AC,Pa_count_MVGDS,Pa_prop_MVGDS,Pa_ratio_PAYRE_vs_MVGDS,Pa_ratio_AC_vs_MVGDS
+locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,18.00000,0.23377,4.00000,0.05195,29.00000,0.37662,0.62069,0.13793,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,19.00000,0.24675,6.00000,0.07792,30.00000,0.38961,0.63333,0.20000,20.00000,0.25974,7.00000,0.09091,32.00000,0.41558,0.62500,0.21875,20.00000,0.25974,5.00000,0.06494,29.00000,0.37662,0.68966,0.17241,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,14.00000,0.21212,2.00000,0.03030,26.00000,0.39394,0.53846,0.07692,19.00000,0.24675,6.00000,0.07792,32.00000,0.41558,0.59375,0.18750
+locus1_sp6_sp6.fasta,24.00000,0.30769,7.00000,0.08974,20.00000,0.25641,1.20000,0.35000,25.00000,0.32051,7.00000,0.08974,20.00000,0.25641,1.25000,0.35000,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,25.00000,0.32051,8.00000,0.10256,21.00000,0.26923,1.19048,0.38095,24.00000,0.30769,7.00000,0.08974,22.00000,0.28205,1.09091,0.31818,24.00000,0.30769,7.00000,0.08974,20.00000,0.25641,1.20000,0.35000,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,24.00000,0.30769,7.00000,0.08974,22.00000,0.28205,1.09091,0.31818
+locus2_sp6_sp6.fasta,18.00000,0.28125,6.00000,0.09375,20.00000,0.31250,0.90000,0.30000,18.00000,0.28125,6.00000,0.09375,21.00000,0.32812,0.85714,0.28571,NA,NA,NA,NA,NA,NA,NA,NA,18.00000,0.28125,6.00000,0.09375,20.00000,0.31250,0.90000,0.30000,NA,NA,NA,NA,NA,NA,NA,NA,17.00000,0.26562,6.00000,0.09375,20.00000,0.31250,0.85000,0.30000,20.00000,0.31250,8.00000,0.12500,19.00000,0.29688,1.05263,0.42105,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,17.00000,0.26562,6.00000,0.09375,20.00000,0.31250,0.85000,0.30000
