+++ galaxy_wrappers/07_MutCount/test-data/OUT_aa/RHKDE.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_count_RHKDE,Ac_prop_RHKDE,Ac_count_RHK,Ac_prop_RHK,Ac_count_DE,Ac_prop_DE,Pu_count_RHKDE,Pu_prop_RHKDE,Pu_count_RHK,Pu_prop_RHK,Pu_count_DE,Pu_prop_DE,Am_count_RHKDE,Am_prop_RHKDE,Am_count_RHK,Am_prop_RHK,Am_count_DE,Am_prop_DE,Ap_count_RHKDE,Ap_prop_RHKDE,Ap_count_RHK,Ap_prop_RHK,Ap_count_DE,Ap_prop_DE,Pf_count_RHKDE,Pf_prop_RHKDE,Pf_count_RHK,Pf_prop_RHK,Pf_count_DE,Pf_prop_DE,Pg_count_RHKDE,Pg_prop_RHKDE,Pg_count_RHK,Pg_prop_RHK,Pg_count_DE,Pg_prop_DE,Ph_count_RHKDE,Ph_prop_RHKDE,Ph_count_RHK,Ph_prop_RHK,Ph_count_DE,Ph_prop_DE,Ps_count_RHKDE,Ps_prop_RHKDE,Ps_count_RHK,Ps_prop_RHK,Ps_count_DE,Ps_prop_DE,Pp_count_RHKDE,Pp_prop_RHKDE,Pp_count_RHK,Pp_prop_RHK,Pp_count_DE,Pp_prop_DE,Pa_count_RHKDE,Pa_prop_RHKDE,Pa_count_RHK,Pa_prop_RHK,Pa_count_DE,Pa_prop_DE,
-locus2_sp6_sp6.fasta,19.00000,0.29688,9.00000,0.14062,10.00000,0.15625,19.00000,0.29688,9.00000,0.14062,10.00000,0.15625,NA,NA,NA,NA,NA,NA,19.00000,0.29688,9.00000,0.14062,10.00000,0.15625,NA,NA,NA,NA,NA,NA,20.00000,0.31250,10.00000,0.15625,10.00000,0.15625,19.00000,0.29688,9.00000,0.14062,10.00000,0.15625,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,20.00000,0.31250,10.00000,0.15625,10.00000,0.15625,
-locus1_sp6_sp6.fasta,28.00000,0.35897,14.00000,0.17949,14.00000,0.17949,30.00000,0.38462,16.00000,0.20513,14.00000,0.17949,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,30.00000,0.38462,16.00000,0.20513,14.00000,0.17949,30.00000,0.38462,16.00000,0.20513,14.00000,0.17949,30.00000,0.38462,16.00000,0.20513,14.00000,0.17949,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,30.00000,0.38462,16.00000,0.20513,14.00000,0.17949,
-locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,14.00000,0.18182,4.00000,0.05195,10.00000,0.12987,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,16.00000,0.20779,5.00000,0.06494,11.00000,0.14286,15.00000,0.19481,4.00000,0.05195,11.00000,0.14286,14.00000,0.18182,2.00000,0.02597,12.00000,0.15584,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,13.00000,0.19697,2.00000,0.03030,11.00000,0.16667,15.00000,0.19481,4.00000,0.05195,11.00000,0.14286,
+LOCUS,Ac_count_RHKDE,Ac_prop_RHKDE,Ac_count_RHK,Ac_prop_RHK,Ac_count_DE,Ac_prop_DE,Pu_count_RHKDE,Pu_prop_RHKDE,Pu_count_RHK,Pu_prop_RHK,Pu_count_DE,Pu_prop_DE,Am_count_RHKDE,Am_prop_RHKDE,Am_count_RHK,Am_prop_RHK,Am_count_DE,Am_prop_DE,Ap_count_RHKDE,Ap_prop_RHKDE,Ap_count_RHK,Ap_prop_RHK,Ap_count_DE,Ap_prop_DE,Pf_count_RHKDE,Pf_prop_RHKDE,Pf_count_RHK,Pf_prop_RHK,Pf_count_DE,Pf_prop_DE,Pg_count_RHKDE,Pg_prop_RHKDE,Pg_count_RHK,Pg_prop_RHK,Pg_count_DE,Pg_prop_DE,Ph_count_RHKDE,Ph_prop_RHKDE,Ph_count_RHK,Ph_prop_RHK,Ph_count_DE,Ph_prop_DE,Ps_count_RHKDE,Ps_prop_RHKDE,Ps_count_RHK,Ps_prop_RHK,Ps_count_DE,Ps_prop_DE,Pp_count_RHKDE,Pp_prop_RHKDE,Pp_count_RHK,Pp_prop_RHK,Pp_count_DE,Pp_prop_DE,Pa_count_RHKDE,Pa_prop_RHKDE,Pa_count_RHK,Pa_prop_RHK,Pa_count_DE,Pa_prop_DE
+locus1_sp8_sp8.fasta,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,14.00000,0.18182,4.00000,0.05195,10.00000,0.12987,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,16.00000,0.20779,5.00000,0.06494,11.00000,0.14286,15.00000,0.19481,4.00000,0.05195,11.00000,0.14286,14.00000,0.18182,2.00000,0.02597,12.00000,0.15584,0.00000,0.00000,0.00000,0.00000,0.00000,0.00000,13.00000,0.19697,2.00000,0.03030,11.00000,0.16667,15.00000,0.19481,4.00000,0.05195,11.00000,0.14286
+locus1_sp6_sp6.fasta,28.00000,0.35897,14.00000,0.17949,14.00000,0.17949,30.00000,0.38462,16.00000,0.20513,14.00000,0.17949,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,30.00000,0.38462,16.00000,0.20513,14.00000,0.17949,30.00000,0.38462,16.00000,0.20513,14.00000,0.17949,30.00000,0.38462,16.00000,0.20513,14.00000,0.17949,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,30.00000,0.38462,16.00000,0.20513,14.00000,0.17949
+locus2_sp6_sp6.fasta,19.00000,0.29688,9.00000,0.14062,10.00000,0.15625,19.00000,0.29688,9.00000,0.14062,10.00000,0.15625,NA,NA,NA,NA,NA,NA,19.00000,0.29688,9.00000,0.14062,10.00000,0.15625,NA,NA,NA,NA,NA,NA,20.00000,0.31250,10.00000,0.15625,10.00000,0.15625,19.00000,0.29688,9.00000,0.14062,10.00000,0.15625,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,20.00000,0.31250,10.00000,0.15625,10.00000,0.15625
