+++ galaxy_wrappers/07_MutCount/test-data/OUT_aa/STNQ.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_count_STNQ,Ac_prop_STNQ,Pu_count_STNQ,Pu_prop_STNQ,Am_count_STNQ,Am_prop_STNQ,Ap_count_STNQ,Ap_prop_STNQ,Pf_count_STNQ,Pf_prop_STNQ,Pg_count_STNQ,Pg_prop_STNQ,Ph_count_STNQ,Ph_prop_STNQ,Ps_count_STNQ,Ps_prop_STNQ,Pp_count_STNQ,Pp_prop_STNQ,Pa_count_STNQ,Pa_prop_STNQ,
-locus2_sp6_sp6.fasta,8.00000,0.12500,8.00000,0.12500,NA,NA,7.00000,0.10938,NA,NA,7.00000,0.10938,8.00000,0.12500,NA,NA,NA,NA,7.00000,0.10938,
-locus1_sp6_sp6.fasta,19.00000,0.24359,15.00000,0.19231,NA,NA,NA,NA,18.00000,0.23077,18.00000,0.23077,18.00000,0.23077,NA,NA,NA,NA,18.00000,0.23077,
-locus1_sp8_sp8.fasta,0.00000,0.00000,28.00000,0.36364,NA,NA,NA,NA,27.00000,0.35065,25.00000,0.32468,27.00000,0.35065,0.00000,0.00000,24.00000,0.36364,26.00000,0.33766,
+LOCUS,Ac_count_STNQ,Ac_prop_STNQ,Pu_count_STNQ,Pu_prop_STNQ,Am_count_STNQ,Am_prop_STNQ,Ap_count_STNQ,Ap_prop_STNQ,Pf_count_STNQ,Pf_prop_STNQ,Pg_count_STNQ,Pg_prop_STNQ,Ph_count_STNQ,Ph_prop_STNQ,Ps_count_STNQ,Ps_prop_STNQ,Pp_count_STNQ,Pp_prop_STNQ,Pa_count_STNQ,Pa_prop_STNQ
+locus1_sp8_sp8.fasta,0.00000,0.00000,28.00000,0.36364,NA,NA,NA,NA,27.00000,0.35065,25.00000,0.32468,27.00000,0.35065,0.00000,0.00000,24.00000,0.36364,26.00000,0.33766
+locus1_sp6_sp6.fasta,19.00000,0.24359,15.00000,0.19231,NA,NA,NA,NA,18.00000,0.23077,18.00000,0.23077,18.00000,0.23077,NA,NA,NA,NA,18.00000,0.23077
+locus2_sp6_sp6.fasta,8.00000,0.12500,8.00000,0.12500,NA,NA,7.00000,0.10938,NA,NA,7.00000,0.10938,8.00000,0.12500,NA,NA,NA,NA,7.00000,0.10938
