+++ galaxy_wrappers/07_MutCount/test-data/OUT_aa/TotalHydratation.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_Total_Hydratation,Pu_Total_Hydratation,Am_Total_Hydratation,Ap_Total_Hydratation,Pf_Total_Hydratation,Pg_Total_Hydratation,Ph_Total_Hydratation,Ps_Total_Hydratation,Pp_Total_Hydratation,Pa_Total_Hydratation,
-locus2_sp6_sp6.fasta,171.50000,171.50000,NA,170.50000,NA,171.00000,171.50000,NA,NA,171.00000,
-locus1_sp6_sp6.fasta,224.00000,224.50000,NA,NA,223.00000,226.00000,226.00000,NA,NA,226.00000,
-locus1_sp8_sp8.fasta,0.00000,188.00000,NA,NA,190.50000,188.50000,192.00000,0.00000,167.50000,189.50000,
+LOCUS,Ac_Total_Hydratation,Pu_Total_Hydratation,Am_Total_Hydratation,Ap_Total_Hydratation,Pf_Total_Hydratation,Pg_Total_Hydratation,Ph_Total_Hydratation,Ps_Total_Hydratation,Pp_Total_Hydratation,Pa_Total_Hydratation
+locus1_sp8_sp8.fasta,0.00000,188.00000,NA,NA,190.50000,188.50000,192.00000,0.00000,167.50000,189.50000
+locus1_sp6_sp6.fasta,224.00000,224.50000,NA,NA,223.00000,226.00000,226.00000,NA,NA,226.00000
+locus2_sp6_sp6.fasta,171.50000,171.50000,NA,170.50000,NA,171.00000,171.50000,NA,NA,171.00000
