+++ galaxy_wrappers/07_MutCount/test-data/OUT_aa/TotalPartialSpecificVolume.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_Total_Partial_Specific_Volume,Pu_Total_Partial_Specific_Volume,Am_Total_Partial_Specific_Volume,Ap_Total_Partial_Specific_Volume,Pf_Total_Partial_Specific_Volume,Pg_Total_Partial_Specific_Volume,Ph_Total_Partial_Specific_Volume,Ps_Total_Partial_Specific_Volume,Pp_Total_Partial_Specific_Volume,Pa_Total_Partial_Specific_Volume,
-locus2_sp6_sp6.fasta,47.40000,47.77000,NA,47.71000,NA,48.17000,47.97000,NA,NA,48.17000,
-locus1_sp6_sp6.fasta,56.53000,56.56000,NA,NA,56.37000,56.43000,56.54000,NA,NA,56.43000,
-locus1_sp8_sp8.fasta,0.00000,56.86000,NA,NA,56.93000,57.09000,56.25000,0.00000,49.09000,57.05000,
+LOCUS,Ac_Total_Partial_Specific_Volume,Pu_Total_Partial_Specific_Volume,Am_Total_Partial_Specific_Volume,Ap_Total_Partial_Specific_Volume,Pf_Total_Partial_Specific_Volume,Pg_Total_Partial_Specific_Volume,Ph_Total_Partial_Specific_Volume,Ps_Total_Partial_Specific_Volume,Pp_Total_Partial_Specific_Volume,Pa_Total_Partial_Specific_Volume
+locus1_sp8_sp8.fasta,0.00000,56.86000,NA,NA,56.93000,57.09000,56.25000,0.00000,49.09000,57.05000
+locus1_sp6_sp6.fasta,56.53000,56.56000,NA,NA,56.37000,56.43000,56.54000,NA,NA,56.43000
+locus2_sp6_sp6.fasta,47.40000,47.77000,NA,47.71000,NA,48.17000,47.97000,NA,NA,48.17000
