+++ galaxy_wrappers/07_MutCount/test-data/OUT_concat/MutCount_concat_log.output
@@ -1,0 +1,96 @@
+
+ ------ Occurences and frequencies of codons, amino-acids, amino-acids types -------
+
+The script counts the number of codons, amino acids, and types of amino acids in sequences,
+as well as the mutation bias from one item to another between 2 sequences.
+
+Counting are then compared to empirical p-values, obtained from bootstrapped sequences obtained from a subset of sequences.
+In the output files, the pvalues indicate the position of the observed data in a distribution of empirical countings obtained from
+a resample of the data. Values above 0.95 indicate a significantly higher counting, values under 0.05 a significantly lower counting.
+    Sequences file : concatenation.fasta
+    Species retained for countings : Ps,Pp,Pu,Ac,Ap,Pf,Pg,Ph,Pi
+
+Processing : reading input file, opening output files, building dictionaries.
+    Warning : countings might be biased and show high differences between species because of high variations of the indels proportions among sequences.
+    Frequences are more representative.
+
+    Indels percent :
+    Ps : 0.174520069808 %
+    Pp : 1.78883071553 %
+    Pu : 21.8586387435 %
+    Ac : 9.90401396161 %
+    Ap : 8.98778359511 %
+    Pf : 15.7504363002 %
+    Pg : 4.31937172775 %
+    Ph : 5.55555555556 %
+    Pi : 8.11518324607 %
+
+Processing : resampling ...
+    Parameters : 200 iterations, 200 codon per resampled sequence, species used : Ps,Pp,Pu,Pf
+
+    10%
+    20%
+    30%
+    40%
+    50%
+    60%
+    70%
+    80%
+    90%
+    100%
+    Done.
+
+Processing : countings....
+
+Countings on Ps
+Countings on Pp
+Countings transitions between Ps and Pp
+Countings on Pu
+Countings transitions between Ps and Pu
+Countings on Ac
+Countings transitions between Ps and Ac
+Countings on Ap
+Countings transitions between Ps and Ap
+Countings on Pf
+Countings transitions between Ps and Pf
+Countings on Pg
+Countings transitions between Ps and Pg
+Countings on Ph
+Countings transitions between Ps and Ph
+Countings on Pi
+Countings transitions between Ps and Pi
+Countings transitions between Pp and Pu
+Countings transitions between Pp and Ac
+Countings transitions between Pp and Ap
+Countings transitions between Pp and Pf
+Countings transitions between Pp and Pg
+Countings transitions between Pp and Ph
+Countings transitions between Pp and Pi
+Countings transitions between Pu and Ac
+Countings transitions between Pu and Ap
+Countings transitions between Pu and Pf
+Countings transitions between Pu and Pg
+Countings transitions between Pu and Ph
+Countings transitions between Pu and Pi
+Countings transitions between Ac and Ap
+Countings transitions between Ac and Pf
+Countings transitions between Ac and Pg
+Countings transitions between Ac and Ph
+Countings transitions between Ac and Pi
+Countings transitions between Ap and Pf
+Countings transitions between Ap and Pg
+Countings transitions between Ap and Ph
+Countings transitions between Ap and Pi
+Countings transitions between Pf and Pg
+Countings transitions between Pf and Ph
+Countings transitions between Pf and Pi
+Countings transitions between Pg and Ph
+Countings transitions between Pg and Pi
+Countings transitions between Ph and Pi
+
+    Done.
+
+Processing : creating dataframes ...
+Writing dataframes to output files ...
+
+Done.
