--- galaxy_wrappers/07_MutCount/test-data/OUT_concat/aa_counts.csv
@@ -1,9 +1,0 @@
-Species,cys,asn,his,ile,ser,gln,lys,met,pro,thr,phe,ala,gly,val,leu,asp,arg,trp,glu,tyr
-Ac,0.036646086437,0.0341565425214,0.0292770364469,0.057458673571,0.115813582952,0.0330611431986,0.0419239195379,0.0306711810396,0.0539733120892,0.0601473809998,0.0480979884485,0.0626369249154,0.0564628560048,0.0762796255726,0.100278828919,0.0371439952201,0.0573590918144,0.0167297351125,0.0295757817168,0.0223063134834
-Ac_pvalue,0.485148514851,0.50495049505,0.386138613861,0.514851485149,0.50495049505,0.50495049505,0.574257425743,0.693069306931,0.544554455446,0.643564356436,0.445544554455,0.50495049505,0.534653465347,0.534653465347,0.594059405941,0.415841584158,0.435643564356,0.435643564356,0.287128712871,0.613861386139
-Am,0.0141342756184,0.0388692579505,0.0212014134276,0.0724381625442,0.0530035335689,0.0388692579505,0.104240282686,0.0247349823322,0.0494699646643,0.0583038869258,0.047703180212,0.0689045936396,0.0530035335689,0.0777385159011,0.0600706713781,0.0371024734982,0.100706713781,0.0159010600707,0.0459363957597,0.017667844523
-Am_pvalue,0.108910891089,0.50495049505,0.386138613861,0.782178217822,0.019801980198,0.50495049505,0.980198019802,0.485148514851,0.386138613861,0.465346534653,0.445544554455,0.50495049505,0.534653465347,0.534653465347,0.128712871287,0.415841584158,0.960396039604,0.435643564356,0.821782178218,0.336633663366
-Ap,0.0356666030898,0.0330917413694,0.0287049399199,0.0551211138661,0.118348273889,0.0323288193782,0.0397673087927,0.032519549876,0.0531184436391,0.0625596032806,0.0534999046348,0.0607476635514,0.0527369826435,0.0748617203891,0.103852756056,0.0360480640854,0.0569330535953,0.0185962235361,0.0280373831776,0.0234598512302
-Ap_pvalue,0.485148514851,0.50495049505,0.386138613861,0.514851485149,0.50495049505,0.50495049505,0.405940594059,0.693069306931,0.544554455446,0.643564356436,0.594059405941,0.50495049505,0.534653465347,0.534653465347,0.594059405941,0.415841584158,0.435643564356,0.435643564356,0.287128712871,0.613861386139
-Pu,0.0220125786164,0.0440251572327,0.0408805031447,0.0660377358491,0.0534591194969,0.0503144654088,0.0283018867925,0.0188679245283,0.0723270440252,0.0377358490566,0.0314465408805,0.0974842767296,0.0566037735849,0.0786163522013,0.0754716981132,0.0660377358491,0.0754716981132,0.00943396226415,0.0440251572327,0.0314465408805
-Pu_pvalue,0.485148514851,0.50495049505,0.386138613861,0.514851485149,0.50495049505,0.50495049505,0.405940594059,0.693069306931,0.544554455446,0.643564356436,0.594059405941,0.50495049505,0.534653465347,0.534653465347,0.594059405941,0.415841584158,0.435643564356,0.435643564356,0.287128712871,0.613861386139
