+++ galaxy_wrappers/07_MutCount/test-data/OUT_concat/aa_freqs.csv
@@ -1,0 +1,28 @@
+Species,A,C,D,E,F,G,H,I,K,L,M,N,P,Q,R,S,T,V,W,Y
+Ac_obs_counts,176.0,48.0,122.0,115.0,96.0,128.0,43.0,107.0,122.0,193.0,42.0,92.0,84.0,62.0,119.0,118.0,129.0,147.0,39.0,83.0
+Ac_obs_freqs,0.0852300242131,0.0232445520581,0.0590799031477,0.0556900726392,0.0464891041162,0.061985472155,0.0208232445521,0.0518159806295,0.0590799031477,0.0934624697337,0.0203389830508,0.0445520581114,0.0406779661017,0.0300242130751,0.0576271186441,0.0571428571429,0.0624697336562,0.071186440678,0.0188861985472,0.0401937046005
+Ac_pvalues,0.990049751244,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.990049751244,0.512437810945,0.990049751244,0.990049751244,0.990049751244,0.0845771144279,0.00497512437811,0.990049751244,0.00497512437811,0.990049751244,0.865671641791,0.990049751244,0.00497512437811,0.00497512437811
+Ap_obs_counts,166.0,47.0,130.0,112.0,101.0,136.0,46.0,106.0,113.0,194.0,44.0,93.0,90.0,58.0,120.0,127.0,125.0,154.0,38.0,86.0
+Ap_obs_freqs,0.0795781399808,0.0225311601151,0.0623202301055,0.0536912751678,0.0484180249281,0.065196548418,0.0220517737296,0.0508149568552,0.0541706615532,0.0930009587728,0.0210930009588,0.0445829338447,0.0431447746884,0.0278044103547,0.0575263662512,0.0608820709492,0.0599232981783,0.0738255033557,0.0182166826462,0.0412272291467
+Ap_pvalues,0.990049751244,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.990049751244,0.512437810945,0.990049751244,0.990049751244,0.990049751244,0.0845771144279,0.00497512437811,0.990049751244,0.00497512437811,0.990049751244,0.810945273632,0.990049751244,0.00497512437811,0.00497512437811
+Pf_obs_counts,143.0,42.0,117.0,112.0,94.0,127.0,46.0,102.0,99.0,189.0,41.0,84.0,88.0,46.0,103.0,129.0,123.0,123.0,40.0,83.0
+Pf_obs_freqs,0.0740548938374,0.0217503883998,0.0605903676851,0.0580010357328,0.0486794407043,0.0657690315898,0.0238218539617,0.0528223718281,0.0512687726567,0.0978767477991,0.0212325220093,0.0435007767996,0.0455722423615,0.0238218539617,0.0533402382185,0.0668047643708,0.063697566028,0.063697566028,0.0207146556189,0.0429829104091
+Pf_pvalues,0.990049751244,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.990049751244,0.512437810945,0.990049751244,0.990049751244,0.990049751244,0.0845771144279,0.00497512437811,0.990049751244,0.00497512437811,0.990049751244,0.865671641791,0.467661691542,0.00497512437811,0.00497512437811
+Pg_obs_counts,154.0,55.0,139.0,123.0,105.0,142.0,50.0,122.0,118.0,204.0,51.0,94.0,91.0,59.0,131.0,130.0,135.0,157.0,42.0,91.0
+Pg_obs_freqs,0.0702234382125,0.0250797993616,0.0633834929321,0.0560875512996,0.0478796169631,0.0647514819881,0.0227998176015,0.0556315549476,0.0538075695394,0.093023255814,0.0232558139535,0.0428636570907,0.0414956680347,0.0269037847697,0.0597355221158,0.0592795257638,0.0615595075239,0.0715914272686,0.0191518467852,0.0414956680347
+Pg_pvalues,0.990049751244,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.990049751244,0.990049751244,0.990049751244,0.990049751244,0.990049751244,0.0845771144279,0.00497512437811,0.990049751244,0.00497512437811,0.990049751244,0.865671641791,0.990049751244,0.00497512437811,0.00497512437811
+Ph_obs_counts,173.0,53.0,123.0,122.0,110.0,137.0,50.0,119.0,107.0,190.0,51.0,102.0,92.0,61.0,123.0,140.0,127.0,150.0,44.0,90.0
+Ph_obs_freqs,0.0799445471349,0.0244916820702,0.0568391866913,0.0563770794824,0.050831792976,0.0633086876155,0.0231053604436,0.0549907578558,0.0494454713494,0.0878003696858,0.0235674676525,0.047134935305,0.0425138632163,0.0281885397412,0.0568391866913,0.0646950092421,0.0586876155268,0.0693160813309,0.0203327171904,0.0415896487985
+Ph_pvalues,0.990049751244,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.990049751244,0.512437810945,0.990049751244,0.990049751244,0.990049751244,0.611940298507,0.00497512437811,0.990049751244,0.00497512437811,0.990049751244,0.810945273632,0.990049751244,0.00497512437811,0.00497512437811
+Pi_obs_counts,159.0,52.0,125.0,116.0,93.0,140.0,41.0,114.0,126.0,175.0,51.0,98.0,97.0,58.0,120.0,113.0,130.0,159.0,42.0,97.0
+Pi_obs_freqs,0.0754985754986,0.0246913580247,0.0593542260209,0.0550807217474,0.0441595441595,0.0664767331434,0.0194681861349,0.0541310541311,0.0598290598291,0.0830959164292,0.0242165242165,0.0465337132004,0.0460588793922,0.0275403608737,0.0569800569801,0.0536562203229,0.0617283950617,0.0754985754986,0.019943019943,0.0460588793922
+Pi_pvalues,0.990049751244,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.990049751244,0.512437810945,0.990049751244,0.970149253731,0.990049751244,0.611940298507,0.00497512437811,0.990049751244,0.00497512437811,0.990049751244,0.865671641791,0.990049751244,0.00497512437811,0.00497512437811
+Pp_obs_counts,163.0,53.0,139.0,124.0,110.0,146.0,50.0,128.0,125.0,207.0,52.0,98.0,96.0,64.0,132.0,128.0,136.0,161.0,45.0,94.0
+Pp_obs_freqs,0.0724122612172,0.0235450910706,0.0617503331853,0.0550866281653,0.0488671701466,0.0648600621946,0.0222123500666,0.0568636161706,0.0555308751666,0.0919591292759,0.0231008440693,0.0435362061306,0.0426477121279,0.0284318080853,0.0586406041759,0.0568636161706,0.0604175921813,0.0715237672146,0.01999111506,0.0417592181253
+Pp_pvalues,0.990049751244,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.990049751244,0.990049751244,0.990049751244,0.990049751244,0.990049751244,0.0845771144279,0.00497512437811,0.990049751244,0.00497512437811,0.990049751244,0.865671641791,0.990049751244,0.00497512437811,0.00497512437811
+Ps_obs_counts,179.0,50.0,141.0,128.0,106.0,148.0,54.0,129.0,127.0,212.0,49.0,95.0,97.0,62.0,129.0,136.0,136.0,166.0,46.0,98.0
+Ps_obs_freqs,0.0782342657343,0.0218531468531,0.0616258741259,0.0559440559441,0.0463286713287,0.0646853146853,0.0236013986014,0.0563811188811,0.055506993007,0.0926573426573,0.0214160839161,0.041520979021,0.0423951048951,0.0270979020979,0.0563811188811,0.0594405594406,0.0594405594406,0.0725524475524,0.0201048951049,0.0428321678322
+Ps_pvalues,0.990049751244,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.990049751244,0.990049751244,0.990049751244,0.990049751244,0.990049751244,0.0845771144279,0.00497512437811,0.990049751244,0.00497512437811,0.990049751244,0.810945273632,0.990049751244,0.00497512437811,0.00497512437811
+Pu_obs_counts,139.0,46.0,107.0,116.0,91.0,107.0,39.0,98.0,105.0,178.0,39.0,87.0,75.0,52.0,97.0,105.0,94.0,114.0,37.0,65.0
+Pu_obs_freqs,0.0776102735902,0.0256839754327,0.0597431602457,0.0647682858738,0.0508096035734,0.0597431602457,0.0217755443886,0.0547180346175,0.0586264656616,0.0993858179788,0.0217755443886,0.0485762144054,0.0418760469012,0.0290340591848,0.0541596873255,0.0586264656616,0.0524846454495,0.0636515912898,0.0206588498046,0.036292573981
+Pu_pvalues,0.990049751244,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.00497512437811,0.990049751244,0.512437810945,0.990049751244,0.990049751244,0.990049751244,0.611940298507,0.00497512437811,0.990049751244,0.00497512437811,0.990049751244,0.00497512437811,0.467661691542,0.00497512437811,0.00497512437811
