--- galaxy_wrappers/07_MutCount/test-data/OUT_concat/aatypes_counts.csv
@@ -1,9 +1,0 @@
-Species,aromatics,polar,unpolar,charged
-Ac,0.0871340370444,0.333798048198,0.383788090022,0.195279824736
-Ac_pvalue,0.455445544554,0.514851485149,0.633663366337,0.336633663366
-Am,0.0812720848057,0.252650176678,0.356890459364,0.309187279152
-Am_pvalue,0.455445544554,0.039603960396,0.425742574257,0.980198019802
-Ap,0.0955559794011,0.335113484646,0.379839786382,0.189490749571
-Ap_pvalue,0.643564356436,0.514851485149,0.544554455446,0.237623762376
-Pu,0.0723270440252,0.279874213836,0.393081761006,0.254716981132
-Pu_pvalue,0.643564356436,0.514851485149,0.544554455446,0.237623762376
