+++ galaxy_wrappers/07_MutCount/test-data/OUT_concat/aatypes_freqs.csv
@@ -1,0 +1,28 @@
+Species,aromatics,charged,polar,unpolar
+Ac_obs_counts,218.0,521.0,533.0,793.0
+Ac_obs_freqs,0.105569007264,0.252300242131,0.258111380145,0.38401937046
+Ac_pvalues,0.00497512437811,0.00497512437811,0.865671641791,0.990049751244
+Ap_obs_counts,225.0,521.0,540.0,800.0
+Ap_obs_freqs,0.107861936721,0.249760306807,0.25886864813,0.383509108341
+Ap_pvalues,0.00497512437811,0.00497512437811,0.865671641791,0.990049751244
+Pf_obs_counts,217.0,477.0,512.0,725.0
+Pf_obs_freqs,0.112377006732,0.247022268255,0.265147591921,0.375453133092
+Pf_pvalues,0.00497512437811,0.00497512437811,0.990049751244,0.990049751244
+Pg_obs_counts,238.0,561.0,564.0,830.0
+Pg_obs_freqs,0.108527131783,0.255813953488,0.257181942544,0.378476972184
+Pg_pvalues,0.00497512437811,0.0199004975124,0.865671641791,0.990049751244
+Ph_obs_counts,244.0,525.0,575.0,820.0
+Ph_obs_freqs,0.112754158965,0.242606284658,0.265711645102,0.378927911275
+Ph_pvalues,0.00497512437811,0.00497512437811,0.990049751244,0.990049751244
+Pi_obs_counts,232.0,528.0,548.0,798.0
+Pi_obs_freqs,0.110161443495,0.250712250712,0.260208926876,0.378917378917
+Pi_pvalues,0.00497512437811,0.00497512437811,0.865671641791,0.990049751244
+Pp_obs_counts,249.0,570.0,575.0,857.0
+Pp_obs_freqs,0.110617503332,0.25322079076,0.255442025766,0.380719680142
+Pp_pvalues,0.00497512437811,0.00497512437811,0.865671641791,0.990049751244
+Ps_obs_counts,250.0,579.0,576.0,883.0
+Ps_obs_freqs,0.109265734266,0.253059440559,0.251748251748,0.385926573427
+Ps_pvalues,0.00497512437811,0.00497512437811,0.865671641791,0.990049751244
+Pu_obs_counts,193.0,464.0,459.0,675.0
+Pu_obs_freqs,0.107761027359,0.259073143495,0.256281407035,0.376884422111
+Pu_pvalues,0.00497512437811,0.0199004975124,0.865671641791,0.990049751244
