--- galaxy_wrappers/07_MutCount/test-data/OUT_concat/aatypes_transitions.csv
@@ -1,13 +1,0 @@
-Species,aromatics>aromatics,polar>aromatics,unpolar>aromatics,charged>aromatics,aromatics>polar,polar>polar,unpolar>polar,charged>polar,aromatics>unpolar,polar>unpolar,unpolar>unpolar,charged>unpolar,aromatics>charged,polar>charged,unpolar>charged,charged>charged
-Am>Ac,0.000000,0.047619,0.021739,0.000000,0.823223,0.000000,0.021739,0.045455,0.990157,-1.000000,0.000000,0.000000,0.000000,-1.000000,0.000000,0.000000
-Am>Ac_pvalue,0.495050,0.613861,0.475248,0.495050,0.712871,0.495050,0.247525,0.297030,0.702970,0.128713,0.495050,0.247525,0.495050,0.128713,0.495050,0.495050
-Ap>Ac,0.000000,0.047101,0.058158,0.044068,0.985186,0.000000,0.156704,0.200000,0.925826,0.058323,0.000000,0.203390,0.443252,-0.264676,-0.544011,0.000000
-Ap>Ac_pvalue,0.495050,0.613861,0.485149,0.792079,0.742574,0.495050,0.544554,0.613861,0.613861,0.495050,0.495050,0.722772,0.772277,0.485149,0.366337,0.495050
-Pu>Ac,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000
-Pu>Ac_pvalue,0.495050,0.495050,0.247525,0.495050,0.495050,0.495050,0.247525,0.247525,0.247525,0.455446,0.495050,0.247525,0.495050,0.613861,0.495050,0.495050
-Ap>Am,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,1.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000
-Ap>Am_pvalue,0.495050,0.495050,0.247525,0.495050,0.495050,0.495050,0.247525,0.247525,0.247525,0.861386,0.495050,0.247525,0.495050,0.613861,0.495050,0.495050
-Pu>Am,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000
-Pu>Am_pvalue,0.495050,0.495050,0.247525,0.495050,0.495050,0.495050,0.247525,0.247525,0.247525,0.455446,0.495050,0.247525,0.495050,0.613861,0.495050,0.495050
-Pu>Ap,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000
-Pu>Ap_pvalue,0.495050,0.495050,0.247525,0.495050,0.495050,0.495050,0.247525,0.247525,0.247525,0.455446,0.495050,0.247525,0.495050,0.613861,0.495050,0.495050
