--- galaxy_wrappers/07_MutCount/test-data/OUT_concat/codons_counts.csv
@@ -1,9 +1,0 @@
-Species,tat,tgt,tct,ttt,tgc,tgg,tac,ttc,tcg,tta,ttg,tcc,tca,gca,gta,gcc,gtc,gcg,gtg,caa,gtt,gct,acc,ggt,cga,cgc,gat,aag,cgg,act,ggg,gga,ggc,gag,aaa,gac,cgt,gaa,ctt,atg,aca,acg,atc,aac,ata,agg,cct,agc,aga,cat,aat,att,ctg,cta,ctc,cac,ccg,agt,cag,cca,ccc
-Ac,0.00975901214897,0.0217088229436,0.0274845648277,0.0231029675363,0.0149372634933,0.0167297351125,0.0125473013344,0.0249950209122,0.0103565026887,0.0130452101175,0.0243975303724,0.0226050587532,0.0248954391555,0.0182234614619,0.0165305715993,0.020812587134,0.0199163513244,0.00527783310098,0.0223063134834,0.0139414459271,0.0175263891655,0.0183230432185,0.0122485560645,0.0159330810595,0.00746863174666,0.00567616012746,0.0181238797052,0.0223063134834,0.00507866958773,0.0144393547102,0.00796654052977,0.0162318263294,0.016331408086,0.0148376817367,0.0196176060546,0.0190201155148,0.00876319458275,0.0147380999801,0.0170284803824,0.0306711810396,0.0253933479387,0.0080661222864,0.0217088229436,0.0183230432185,0.0185222067317,0.0139414459271,0.016331408086,0.0159330810595,0.0164309898427,0.0156343357897,0.0158334993029,0.0172276438956,0.0247958573989,0.00637323242382,0.0146385182235,0.0136427006572,0.00647281418044,0.0145389364668,0.0191196972715,0.0231029675363,0.0080661222864
-Ac_pvalue,0.336633663366,0.594059405941,0.50495049505,0.663366336634,0.60396039604,0.435643564356,0.673267326733,0.465346534653,0.80198019802,0.712871287129,0.50495049505,0.594059405941,0.534653465347,0.376237623762,0.544554455446,0.623762376238,0.336633663366,0.544554455446,0.633663366337,0.554455445545,0.514851485149,0.405940594059,0.594059405941,0.613861386139,0.455445544554,0.405940594059,0.376237623762,0.594059405941,0.554455445545,0.60396039604,0.465346534653,0.594059405941,0.534653465347,0.50495049505,0.485148514851,0.376237623762,0.376237623762,0.50495049505,0.455445544554,0.693069306931,0.554455445545,0.405940594059,0.594059405941,0.346534653465,0.49504950495,0.594059405941,0.534653465347,0.465346534653,0.544554455446,0.425742574257,0.554455445545,0.465346534653,0.574257425743,0.574257425743,0.514851485149,0.564356435644,0.445544554455,0.623762376238,0.455445544554,0.584158415842,0.405940594059
-Am,0.00530035335689,0.00883392226148,0.00883392226148,0.0229681978799,0.00530035335689,0.0159010600707,0.0123674911661,0.0247349823322,0.0106007067138,0.0017667844523,0.0106007067138,0.0141342756184,0.00706713780919,0.00706713780919,0.00530035335689,0.0353356890459,0.0282685512367,0.017667844523,0.030035335689,0.00530035335689,0.0141342756184,0.00883392226148,0.0318021201413,0.0123674911661,0.0318021201413,0.0282685512367,0.00883392226148,0.0848056537102,0.00706713780919,0.00883392226148,0.0106007067138,0.0123674911661,0.017667844523,0.0353356890459,0.0194346289753,0.0282685512367,0.00883392226148,0.0106007067138,0.0,0.0247349823322,0.00706713780919,0.0106007067138,0.047703180212,0.030035335689,0.00883392226148,0.017667844523,0.00530035335689,0.0106007067138,0.00706713780919,0.00353356890459,0.00883392226148,0.0159010600707,0.0318021201413,0.0,0.0159010600707,0.017667844523,0.0159010600707,0.0017667844523,0.0335689045936,0.017667844523,0.0106007067138
-Am_pvalue,0.336633663366,0.0693069306931,0.0594059405941,0.663366336634,0.267326732673,0.435643564356,0.673267326733,0.465346534653,0.80198019802,0.29702970297,0.267326732673,0.336633663366,0.049504950495,0.108910891089,0.257425742574,0.871287128713,0.663366336634,0.851485148515,0.871287128713,0.227722772277,0.514851485149,0.168316831683,0.940594059406,0.613861386139,0.980198019802,0.871287128713,0.0990099009901,0.980198019802,0.554455445545,0.316831683168,0.831683168317,0.594059405941,0.534653465347,0.920792079208,0.485148514851,0.643564356436,0.376237623762,0.50495049505,0.128712871287,0.485148514851,0.0891089108911,0.851485148515,0.891089108911,0.871287128713,0.138613861386,0.594059405941,0.227722772277,0.465346534653,0.237623762376,0.207920792079,0.217821782178,0.465346534653,0.782178217822,0.49504950495,0.514851485149,0.564356435644,0.851485148515,0.257425742574,0.851485148515,0.376237623762,0.811881188119
-Ap,0.00905969864581,0.0208849895098,0.0277512874309,0.0241274079725,0.01478161358,0.0185962235361,0.0144001525844,0.0293724966622,0.0101087163837,0.0119206561129,0.0277512874309,0.0229830249857,0.0262254434484,0.016784283807,0.0142094220866,0.0213618157543,0.0199313370208,0.00505435819187,0.0230783902346,0.0120160213618,0.0176425710471,0.0175472057982,0.0146862483311,0.0133511348465,0.00629410642762,0.00553118443639,0.0180240320427,0.0220293724967,0.00553118443639,0.0133511348465,0.00715239366775,0.01478161358,0.0174518405493,0.0135418653443,0.017737936296,0.0180240320427,0.00934579439252,0.0144955178333,0.017737936296,0.032519549876,0.0249856952127,0.00953652489033,0.022696929239,0.0179286667938,0.015735266069,0.0143047873355,0.0162120923136,0.0176425710471,0.0159259965669,0.0136372305932,0.0151630745756,0.0166889185581,0.0248903299638,0.00629410642762,0.0152584398245,0.0150677093267,0.00600801068091,0.0136372305932,0.0203127980164,0.0218386419989,0.00905969864581
-Ap_pvalue,0.336633663366,0.594059405941,0.50495049505,0.663366336634,0.60396039604,0.435643564356,0.673267326733,0.465346534653,0.80198019802,0.712871287129,0.50495049505,0.594059405941,0.534653465347,0.376237623762,0.544554455446,0.623762376238,0.336633663366,0.544554455446,0.633663366337,0.554455445545,0.514851485149,0.405940594059,0.594059405941,0.613861386139,0.455445544554,0.405940594059,0.376237623762,0.594059405941,0.554455445545,0.60396039604,0.465346534653,0.594059405941,0.534653465347,0.50495049505,0.485148514851,0.376237623762,0.376237623762,0.50495049505,0.455445544554,0.693069306931,0.554455445545,0.405940594059,0.594059405941,0.346534653465,0.49504950495,0.594059405941,0.534653465347,0.465346534653,0.544554455446,0.425742574257,0.554455445545,0.465346534653,0.574257425743,0.574257425743,0.514851485149,0.564356435644,0.445544554455,0.623762376238,0.663366336634,0.584158415842,0.405940594059
-Pu,0.0188679245283,0.0188679245283,0.0125786163522,0.0125786163522,0.00314465408805,0.00943396226415,0.0125786163522,0.0188679245283,0.0062893081761,0.00314465408805,0.00943396226415,0.0188679245283,0.0,0.0251572327044,0.0251572327044,0.0220125786164,0.0157232704403,0.0125786163522,0.0157232704403,0.0377358490566,0.0220125786164,0.0377358490566,0.00943396226415,0.00314465408805,0.0220125786164,0.0188679245283,0.0283018867925,0.0062893081761,0.00943396226415,0.0157232704403,0.0220125786164,0.00943396226415,0.0220125786164,0.0188679245283,0.0220125786164,0.0377358490566,0.0157232704403,0.0251572327044,0.0188679245283,0.0188679245283,0.00943396226415,0.00314465408805,0.0157232704403,0.0157232704403,0.0251572327044,0.00314465408805,0.0157232704403,0.00943396226415,0.0062893081761,0.0220125786164,0.0283018867925,0.0251572327044,0.0188679245283,0.00314465408805,0.0220125786164,0.0188679245283,0.00943396226415,0.0062893081761,0.0125786163522,0.0314465408805,0.0157232704403
-Pu_pvalue,0.336633663366,0.594059405941,0.50495049505,0.663366336634,0.60396039604,0.435643564356,0.673267326733,0.465346534653,0.80198019802,0.712871287129,0.50495049505,0.594059405941,0.534653465347,0.376237623762,0.544554455446,0.623762376238,0.336633663366,0.544554455446,0.633663366337,0.554455445545,0.514851485149,0.405940594059,0.594059405941,0.613861386139,0.455445544554,0.405940594059,0.376237623762,0.594059405941,0.554455445545,0.60396039604,0.465346534653,0.594059405941,0.534653465347,0.50495049505,0.485148514851,0.376237623762,0.376237623762,0.50495049505,0.455445544554,0.693069306931,0.554455445545,0.405940594059,0.594059405941,0.346534653465,0.49504950495,0.594059405941,0.534653465347,0.465346534653,0.544554455446,0.425742574257,0.554455445545,0.465346534653,0.574257425743,0.574257425743,0.514851485149,0.564356435644,0.445544554455,0.623762376238,0.663366336634,0.584158415842,0.405940594059
