--- galaxy_wrappers/07_MutCount/test-data/OUT_concat/counts.txt
@@ -1,66 +1,0 @@
- background: 100 100 + all
-
-
-counting of Ac
-
-occurence of codons
-	tat	tgt	tct	ttt	tgc	tgg	tac	ttc	tcg	tta	ttg	tcc	tca	gca	gta	gcc	gtc	gcg	gtg	caa	gtt	gct	acc	ggt	cga	cgc	gat	aag	cgg	act	ggg	gga	ggc	gag	aaa	gac	cgt	gaa	ctt	atg	aca	acg	atc	aac	ata	agg	cct	agc	aga	cat	aat	att	ctg	cta	ctc	cac	ccg	agt	cag	cca	ccc
-	0.009759	0.021709	0.027485	0.023103	0.014937	0.016730	0.012547	0.024995	0.010357	0.013045	0.024398	0.022605	0.024895	0.018223	0.016531	0.020813	0.019916	0.005278	0.022306	0.013941	0.017526	0.018323	0.012249	0.015933	0.007469	0.005676	0.018124	0.022306	0.005079	0.014439	0.007967	0.016232	0.016331	0.014838	0.019618	0.019020	0.008763	0.014738	0.017028	0.030671	0.025393	0.008066	0.021709	0.018323	0.018522	0.013941	0.016331	0.015933	0.016431	0.015634	0.015833	0.017228	0.024796	0.006373	0.014639	0.013643	0.006473	0.014539	0.019120	0.023103	0.008066
-pvalues	0.336634	0.594059	0.504950	0.663366	0.603960	0.435644	0.673267	0.465347	0.801980	0.712871	0.504950	0.594059	0.534653	0.376238	0.544554	0.623762	0.336634	0.544554	0.633663	0.554455	0.514851	0.405941	0.594059	0.613861	0.455446	0.405941	0.376238	0.594059	0.554455	0.603960	0.465347	0.594059	0.534653	0.504950	0.485149	0.376238	0.376238	0.504950	0.455446	0.693069	0.554455	0.405941	0.594059	0.346535	0.495050	0.594059	0.534653	0.465347	0.544554	0.425743	0.554455	0.465347	0.574257	0.574257	0.514851	0.564356	0.445545	0.623762	0.455446	0.584158	0.405941
-
-occurence of amino_acids
-	cys	asn	his	ile	ser	gln	lys	met	pro	thr	phe	ala	gly	val	leu	asp	arg	trp	glu	tyr
-	0.036646	0.034157	0.029277	0.057459	0.115814	0.033061	0.041924	0.030671	0.053973	0.060147	0.048098	0.062637	0.056463	0.076280	0.100279	0.037144	0.057359	0.016730	0.029576	0.022306
-pvalues	0.485149	0.504950	0.386139	0.514851	0.504950	0.504950	0.574257	0.693069	0.544554	0.643564	0.445545	0.504950	0.534653	0.534653	0.594059	0.415842	0.435644	0.435644	0.287129	0.613861
-
-occurence of amino_acid types
-	aromatics	polar	unpolar	charged
-	0.087134	0.333798	0.383788	0.195280
-pvalues	0.455446	0.514851	0.633663	0.336634
-
-GC3	GC12	IVYWREL	EK/QH	PAYRE/SDGM	Purine_load	CvP
-0.493726	0.464001	0.359988	1.146965	0.940689	-40.496581	-0.138518
-
-
-counting of Am
-
-occurence of codons
-	tat	tgt	tct	ttt	tgc	tgg	tac	ttc	tcg	tta	ttg	tcc	tca	gca	gta	gcc	gtc	gcg	gtg	caa	gtt	gct	acc	ggt	cga	cgc	gat	aag	cgg	act	ggg	gga	ggc	gag	aaa	gac	cgt	gaa	ctt	atg	aca	acg	atc	aac	ata	agg	cct	agc	aga	cat	aat	att	ctg	cta	ctc	cac	ccg	agt	cag	cca	ccc
-	0.005300	0.008834	0.008834	0.022968	0.005300	0.015901	0.012367	0.024735	0.010601	0.001767	0.010601	0.014134	0.007067	0.007067	0.005300	0.035336	0.028269	0.017668	0.030035	0.005300	0.014134	0.008834	0.031802	0.012367	0.031802	0.028269	0.008834	0.084806	0.007067	0.008834	0.010601	0.012367	0.017668	0.035336	0.019435	0.028269	0.008834	0.010601	0.000000	0.024735	0.007067	0.010601	0.047703	0.030035	0.008834	0.017668	0.005300	0.010601	0.007067	0.003534	0.008834	0.015901	0.031802	0.000000	0.015901	0.017668	0.015901	0.001767	0.033569	0.017668	0.010601
-pvalues	0.336634	0.069307	0.059406	0.663366	0.267327	0.435644	0.673267	0.465347	0.801980	0.297030	0.267327	0.336634	0.049505	0.108911	0.257426	0.871287	0.663366	0.851485	0.871287	0.227723	0.514851	0.168317	0.940594	0.613861	0.980198	0.871287	0.099010	0.980198	0.554455	0.316832	0.831683	0.594059	0.534653	0.920792	0.485149	0.643564	0.376238	0.504950	0.128713	0.485149	0.089109	0.851485	0.891089	0.871287	0.138614	0.594059	0.227723	0.465347	0.237624	0.207921	0.217822	0.465347	0.782178	0.495050	0.514851	0.564356	0.851485	0.257426	0.851485	0.376238	0.811881
-
-occurence of amino_acids
-	cys	asn	his	ile	ser	gln	lys	met	pro	thr	phe	ala	gly	val	leu	asp	arg	trp	glu	tyr
-	0.014134	0.038869	0.021201	0.072438	0.053004	0.038869	0.104240	0.024735	0.049470	0.058304	0.047703	0.068905	0.053004	0.077739	0.060071	0.037102	0.100707	0.015901	0.045936	0.017668
-pvalues	0.108911	0.504950	0.386139	0.782178	0.019802	0.504950	0.980198	0.485149	0.386139	0.465347	0.445545	0.504950	0.534653	0.534653	0.128713	0.415842	0.960396	0.435644	0.821782	0.336634
-
-occurence of amino_acid types
-	aromatics	polar	unpolar	charged
-	0.081272	0.252650	0.356890	0.309187
-pvalues	0.455446	0.039604	0.425743	0.980198
-
-GC3	GC12	IVYWREL	EK/QH	PAYRE/SDGM	Purine_load	CvP
-0.715548	0.464664	0.390459	2.500000	1.684211	77.738516	0.056537
-
-
-counting of Ap
-
-occurence of codons
-	tat	tgt	tct	ttt	tgc	tgg	tac	ttc	tcg	tta	ttg	tcc	tca	gca	gta	gcc	gtc	gcg	gtg	caa	gtt	gct	acc	ggt	cga	cgc	gat	aag	cgg	act	ggg	gga	ggc	gag	aaa	gac	cgt	gaa	ctt	atg	aca	acg	atc	aac	ata	agg	cct	agc	aga	cat	aat	att	ctg	cta	ctc	cac	ccg	agt	cag	cca	ccc
-	0.009060	0.020885	0.027751	0.024127	0.014782	0.018596	0.014400	0.029372	0.010109	0.011921	0.027751	0.022983	0.026225	0.016784	0.014209	0.021362	0.019931	0.005054	0.023078	0.012016	0.017643	0.017547	0.014686	0.013351	0.006294	0.005531	0.018024	0.022029	0.005531	0.013351	0.007152	0.014782	0.017452	0.013542	0.017738	0.018024	0.009346	0.014496	0.017738	0.032520	0.024986	0.009537	0.022697	0.017929	0.015735	0.014305	0.016212	0.017643	0.015926	0.013637	0.015163	0.016689	0.024890	0.006294	0.015258	0.015068	0.006008	0.013637	0.020313	0.021839	0.009060
-pvalues	0.336634	0.594059	0.504950	0.663366	0.603960	0.435644	0.673267	0.465347	0.801980	0.712871	0.504950	0.594059	0.534653	0.376238	0.544554	0.623762	0.336634	0.544554	0.633663	0.554455	0.514851	0.405941	0.594059	0.613861	0.455446	0.405941	0.376238	0.594059	0.554455	0.603960	0.465347	0.594059	0.534653	0.504950	0.485149	0.376238	0.376238	0.504950	0.455446	0.693069	0.554455	0.405941	0.594059	0.346535	0.495050	0.594059	0.534653	0.465347	0.544554	0.425743	0.554455	0.465347	0.574257	0.574257	0.514851	0.564356	0.445545	0.623762	0.663366	0.584158	0.405941
-
-occurence of amino_acids
-	cys	asn	his	ile	ser	gln	lys	met	pro	thr	phe	ala	gly	val	leu	asp	arg	trp	glu	tyr
-	0.035667	0.033092	0.028705	0.055121	0.118348	0.032329	0.039767	0.032520	0.053118	0.062560	0.053500	0.060748	0.052737	0.074862	0.103853	0.036048	0.056933	0.018596	0.028037	0.023460
-pvalues	0.485149	0.504950	0.386139	0.514851	0.504950	0.504950	0.405941	0.693069	0.544554	0.643564	0.594059	0.504950	0.534653	0.534653	0.594059	0.415842	0.435644	0.435644	0.287129	0.613861
-
-occurence of amino_acid types
-	aromatics	polar	unpolar	charged
-	0.095556	0.335113	0.379840	0.189491
-pvalues	0.643564	0.514851	0.544554	0.237624
-
-GC3	GC12	IVYWREL	EK/QH	PAYRE/SDGM	Purine_load	CvP
-0.516594	0.458087	0.360862	1.110937	0.927577	-57.791341	-0.145623
-
-
