+++ galaxy_wrappers/07_MutCount/test-data/OUT_concat/gc_and_others_freqs.csv
@@ -1,0 +1,10 @@
+Species,GC3,GC12,IVYWREL,EKQH,PAYRESDGM,purineload,CvP
+Ac,0.434991273997,0.417102966841,0.35034904014,2.25714285714,1.40731707317,42.6119837115,-0.00581113801453
+Ap,0.468586387435,0.425392670157,0.353403141361,2.16346153846,1.31350114416,31.1227457824,-0.00910834132311
+Pf,0.428010471204,0.387216404887,0.328097731239,2.29347826087,1.27777777778,26.6143106457,-0.0181253236665
+Pg,0.452443280977,0.436954624782,0.379581151832,2.21100917431,1.27705627706,41.4485165794,-0.00136798905609
+Ph,0.460732984293,0.434554973822,0.365619546248,2.06306306306,1.33037694013,28.2140779523,-0.0231053604436
+Pi,0.448516579407,0.421465968586,0.35907504363,2.44444444444,1.37296037296,44.502617801,-0.00949667616334
+Pp,0.458987783595,0.446989528796,0.388743455497,2.18421052632,1.30967741935,46.3932518906,-0.00222123500666
+Ps,0.457242582897,0.458115183246,0.396160558464,2.19827586207,1.33122362869,38.3944153578,0.00131118881119
+Pu,0.422338568935,0.355366492147,0.307591623037,2.42857142857,1.37430167598,31.2681791739,0.00279173646008
