--- galaxy_wrappers/07_MutCount/test-data/OUT_concat/gc_counts.csv
@@ -1,5 +1,0 @@
-Species,GC3,GC12,IVYWREL,EKQH,PAYRESDGM,purineload,CvP
-Ac,0.493726349333,0.464001194981,0.359988050189,1.14696485623,0.940688510991,-40.4965810264,-0.138518223461
-Am,0.71554770318,0.464664310954,0.390459363958,2.5,1.68421052632,77.7385159011,0.0565371024735
-Ap,0.516593553309,0.458086973107,0.36086210185,1.1109375,0.927576601671,-57.7913408354,-0.145622735075
-Pu,0.452830188679,0.529874213836,0.380503144654,0.793103448276,1.64516129032,-20.964360587,-0.0251572327044
