+++ galaxy_wrappers/07_MutCount/test-data/OUT_nuc/Purine_Load_Indice.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_TOTAL,Ac_DIFF_GC,Ac_DIFF_AT,Ac_PLI_GC1000,Ac_PLI_AT1000,Pu_TOTAL,Pu_DIFF_GC,Pu_DIFF_AT,Pu_PLI_GC1000,Pu_PLI_AT1000,Am_TOTAL,Am_DIFF_GC,Am_DIFF_AT,Am_PLI_GC1000,Am_PLI_AT1000,Ap_TOTAL,Ap_DIFF_GC,Ap_DIFF_AT,Ap_PLI_GC1000,Ap_PLI_AT1000,Pf_TOTAL,Pf_DIFF_GC,Pf_DIFF_AT,Pf_PLI_GC1000,Pf_PLI_AT1000,Pg_TOTAL,Pg_DIFF_GC,Pg_DIFF_AT,Pg_PLI_GC1000,Pg_PLI_AT1000,Ph_TOTAL,Ph_DIFF_GC,Ph_DIFF_AT,Ph_PLI_GC1000,Ph_PLI_AT1000,Ps_TOTAL,Ps_DIFF_GC,Ps_DIFF_AT,Ps_PLI_GC1000,Ps_PLI_AT1000,Pp_TOTAL,Pp_DIFF_GC,Pp_DIFF_AT,Pp_PLI_GC1000,Pp_PLI_AT1000,Pa_TOTAL,Pa_DIFF_GC,Pa_DIFF_AT,Pa_PLI_GC1000,Pa_PLI_AT1000,
-locus2_sp6_sp6.fasta,192,12,12,62.50000,62.50000,192,12,4,62.50000,20.83333,NA,NA,NA,NA,NA,192,11,11,57.29167,57.29167,NA,NA,NA,NA,NA,192,16,10,83.33333,52.08333,192,10,20,52.08333,104.16667,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,192,14,12,72.91667,62.50000,
-locus1_sp6_sp6.fasta,234,2,28,8.54701,119.65812,234,6,20,25.64103,85.47009,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,234,18,20,76.92308,85.47009,234,0,24,0.00000,102.56410,234,8,16,34.18803,68.37607,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,234,-1,23,-4.27350,98.29060,
-locus1_sp8_sp8.fasta,198,6,20,30.30303,101.01010,231,4,19,17.31602,82.25108,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,231,6,21,25.97403,90.90909,231,15,8,64.93506,34.63203,231,19,10,82.25108,43.29004,0,0,0,0.00000,0.00000,198,10,14,50.50505,70.70707,231,13,10,56.27706,43.29004,
+LOCUS,Ac_TOTAL,Ac_DIFF_GC,Ac_DIFF_AT,Ac_PLI_GC1000,Ac_PLI_AT1000,Pu_TOTAL,Pu_DIFF_GC,Pu_DIFF_AT,Pu_PLI_GC1000,Pu_PLI_AT1000,Am_TOTAL,Am_DIFF_GC,Am_DIFF_AT,Am_PLI_GC1000,Am_PLI_AT1000,Ap_TOTAL,Ap_DIFF_GC,Ap_DIFF_AT,Ap_PLI_GC1000,Ap_PLI_AT1000,Pf_TOTAL,Pf_DIFF_GC,Pf_DIFF_AT,Pf_PLI_GC1000,Pf_PLI_AT1000,Pg_TOTAL,Pg_DIFF_GC,Pg_DIFF_AT,Pg_PLI_GC1000,Pg_PLI_AT1000,Ph_TOTAL,Ph_DIFF_GC,Ph_DIFF_AT,Ph_PLI_GC1000,Ph_PLI_AT1000,Ps_TOTAL,Ps_DIFF_GC,Ps_DIFF_AT,Ps_PLI_GC1000,Ps_PLI_AT1000,Pp_TOTAL,Pp_DIFF_GC,Pp_DIFF_AT,Pp_PLI_GC1000,Pp_PLI_AT1000,Pa_TOTAL,Pa_DIFF_GC,Pa_DIFF_AT,Pa_PLI_GC1000,Pa_PLI_AT1000
+locus1_sp8_sp8.fasta,198,6,20,30.30303,101.01010,231,4,19,17.31602,82.25108,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,231,6,21,25.97403,90.90909,231,15,8,64.93506,34.63203,231,19,10,82.25108,43.29004,0,0,0,0.00000,0.00000,198,10,14,50.50505,70.70707,231,13,10,56.27706,43.29004
+locus1_sp6_sp6.fasta,234,2,28,8.54701,119.65812,234,6,20,25.64103,85.47009,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,234,18,20,76.92308,85.47009,234,0,24,0.00000,102.56410,234,8,16,34.18803,68.37607,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,234,-1,23,-4.27350,98.29060
+locus2_sp6_sp6.fasta,192,12,12,62.50000,62.50000,192,12,4,62.50000,20.83333,NA,NA,NA,NA,NA,192,11,11,57.29167,57.29167,NA,NA,NA,NA,NA,192,16,10,83.33333,52.08333,192,10,20,52.08333,104.16667,NA,NA,NA,NA,NA,NA,NA,NA,NA,NA,192,14,12,72.91667,62.50000
