+++ galaxy_wrappers/07_MutCount/test-data/OUT_nuc/nuc_compositions.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_prop_A,Ac_prop_T,Ac_prop_C,Ac_prop_G,Pu_prop_A,Pu_prop_T,Pu_prop_C,Pu_prop_G,Am_prop_A,Am_prop_T,Am_prop_C,Am_prop_G,Ap_prop_A,Ap_prop_T,Ap_prop_C,Ap_prop_G,Pf_prop_A,Pf_prop_T,Pf_prop_C,Pf_prop_G,Pg_prop_A,Pg_prop_T,Pg_prop_C,Pg_prop_G,Ph_prop_A,Ph_prop_T,Ph_prop_C,Ph_prop_G,Ps_prop_A,Ps_prop_T,Ps_prop_C,Ps_prop_G,Pp_prop_A,Pp_prop_T,Pp_prop_C,Pp_prop_G,Pa_prop_A,Pa_prop_T,Pa_prop_C,Pa_prop_G,
-locus2_sp6_sp6.fasta,0.30208,0.23958,0.19792,0.26042,0.29688,0.27604,0.18229,0.24479,NA,NA,NA,NA,0.30208,0.24479,0.19792,0.25521,NA,NA,NA,NA,0.31250,0.26042,0.17188,0.25521,0.32292,0.21875,0.20312,0.25521,NA,NA,NA,NA,NA,NA,NA,NA,0.31771,0.25521,0.17708,0.25000,
-locus1_sp6_sp6.fasta,0.33761,0.21795,0.21795,0.22650,0.28632,0.20085,0.24359,0.26923,NA,NA,NA,NA,NA,NA,NA,NA,0.29915,0.21368,0.20513,0.28205,0.31624,0.21368,0.23504,0.23504,0.30342,0.23504,0.21368,0.24786,NA,NA,NA,NA,NA,NA,NA,NA,0.31624,0.21795,0.23504,0.23077,
-locus1_sp8_sp8.fasta,0.33333,0.23232,0.20202,0.23232,0.32468,0.24242,0.20779,0.22511,NA,NA,NA,NA,NA,NA,NA,NA,0.32900,0.23810,0.20346,0.22944,0.29437,0.25974,0.19048,0.25541,0.32035,0.27706,0.16017,0.24242,0.00000,0.00000,0.00000,0.00000,0.34343,0.27273,0.16667,0.21717,0.29870,0.25541,0.19481,0.25108,
+LOCUS,Ac_prop_A,Ac_prop_T,Ac_prop_C,Ac_prop_G,Pu_prop_A,Pu_prop_T,Pu_prop_C,Pu_prop_G,Am_prop_A,Am_prop_T,Am_prop_C,Am_prop_G,Ap_prop_A,Ap_prop_T,Ap_prop_C,Ap_prop_G,Pf_prop_A,Pf_prop_T,Pf_prop_C,Pf_prop_G,Pg_prop_A,Pg_prop_T,Pg_prop_C,Pg_prop_G,Ph_prop_A,Ph_prop_T,Ph_prop_C,Ph_prop_G,Ps_prop_A,Ps_prop_T,Ps_prop_C,Ps_prop_G,Pp_prop_A,Pp_prop_T,Pp_prop_C,Pp_prop_G,Pa_prop_A,Pa_prop_T,Pa_prop_C,Pa_prop_G
+locus1_sp8_sp8.fasta,0.33333,0.23232,0.20202,0.23232,0.32468,0.24242,0.20779,0.22511,NA,NA,NA,NA,NA,NA,NA,NA,0.32900,0.23810,0.20346,0.22944,0.29437,0.25974,0.19048,0.25541,0.32035,0.27706,0.16017,0.24242,0.00000,0.00000,0.00000,0.00000,0.34343,0.27273,0.16667,0.21717,0.29870,0.25541,0.19481,0.25108
+locus1_sp6_sp6.fasta,0.33761,0.21795,0.21795,0.22650,0.28632,0.20085,0.24359,0.26923,NA,NA,NA,NA,NA,NA,NA,NA,0.29915,0.21368,0.20513,0.28205,0.31624,0.21368,0.23504,0.23504,0.30342,0.23504,0.21368,0.24786,NA,NA,NA,NA,NA,NA,NA,NA,0.31624,0.21795,0.23504,0.23077
+locus2_sp6_sp6.fasta,0.30208,0.23958,0.19792,0.26042,0.29688,0.27604,0.18229,0.24479,NA,NA,NA,NA,0.30208,0.24479,0.19792,0.25521,NA,NA,NA,NA,0.31250,0.26042,0.17188,0.25521,0.32292,0.21875,0.20312,0.25521,NA,NA,NA,NA,NA,NA,NA,NA,0.31771,0.25521,0.17708,0.25000
