+++ galaxy_wrappers/07_MutCount/test-data/OUT_nuc/percent_purine.csv
@@ -1,4 +1,4 @@
-LOCUS,Ac_percent_purine,Pu_percent_purine,Am_percent_purine,Ap_percent_purine,Pf_percent_purine,Pg_percent_purine,Ph_percent_purine,Ps_percent_purine,Pp_percent_purine,Pa_percent_purine,
-locus2_sp6_sp6.fasta,56.25000,54.16667,NA,55.72917,NA,56.77083,57.81250,NA,NA,56.77083,
-locus1_sp6_sp6.fasta,56.41026,55.55556,NA,NA,58.11966,55.12821,55.12821,NA,NA,54.70085,
-locus1_sp8_sp8.fasta,56.56566,54.97835,NA,NA,55.84416,54.97835,56.27706,0.00000,56.06061,54.97835,
+LOCUS,Ac_percent_purine,Pu_percent_purine,Am_percent_purine,Ap_percent_purine,Pf_percent_purine,Pg_percent_purine,Ph_percent_purine,Ps_percent_purine,Pp_percent_purine,Pa_percent_purine
+locus1_sp8_sp8.fasta,56.56566,54.97835,NA,NA,55.84416,54.97835,56.27706,0.00000,56.06061,54.97835
+locus1_sp6_sp6.fasta,56.41026,55.55556,NA,NA,58.11966,55.12821,55.12821,NA,NA,54.70085
+locus2_sp6_sp6.fasta,56.25000,54.16667,NA,55.72917,NA,56.77083,57.81250,NA,NA,56.77083
